% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_interpret_heterogeneity.R
\name{analyze_concordance}
\alias{analyze_concordance}
\title{Analyze concordance between methods}
\usage{
analyze_concordance(evidence_summary)
}
\description{
Analyze concordance between methods
}
\keyword{internal}
