% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_policy_tree.R,
%   R/margot_recalculate_policy_trees.R
\name{apply_covariate_exclusions}
\alias{apply_covariate_exclusions}
\title{Apply covariate exclusions (copied from margot_recalculate_policy_trees.R)}
\usage{
apply_covariate_exclusions(covariates, exclude_patterns, verbose = TRUE)

apply_covariate_exclusions(covariates, exclude_patterns, verbose = TRUE)
}
\description{
Apply covariate exclusions (copied from margot_recalculate_policy_trees.R)

Apply covariate exclusions
}
\keyword{internal}
