% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_qini_simple_baseline.R
\name{average_gain.qini_simple_baseline}
\alias{average_gain.qini_simple_baseline}
\title{Compute Average Gain for Simple Baseline}
\usage{
average_gain.qini_simple_baseline(object, spend, ...)
}
\arguments{
\item{object}{A qini_simple_baseline object}

\item{spend}{The spend level (proportion treated)}

\item{...}{Additional arguments (ignored)}
}
\value{
A list with estimate and std.err
}
\description{
Compute Average Gain for Simple Baseline
}
\keyword{internal}
