% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate_methods_causal_interventions.R
\name{boilerplate_methods_causal_interventions}
\alias{boilerplate_methods_causal_interventions}
\title{Generate Causal Interventions and Contrasts Section for Methods}
\usage{
boilerplate_methods_causal_interventions(
  causal_interventions,
  exposure_var,
  contrasts = "pairwise",
  null_intervention = NULL
)
}
\arguments{
\item{causal_interventions}{A character vector specifying the causal interventions.
Use "exposure_var" as a placeholder for the exposure variable name.}

\item{exposure_var}{A character string specifying the name of the exposure variable.}

\item{contrasts}{A character string specifying the type of contrasts.
Options are "pairwise", "null", or "custom". Default is "pairwise".}

\item{null_intervention}{A character string specifying the null intervention
when using "null" contrasts. Default is NULL.}
}
\value{
A character string containing the markdown-formatted section on
  causal interventions and contrasts.
}
\description{
This function generates a markdown-formatted section describing the causal
interventions and contrasts for a causal inference study. It details the
interventions considered and the approach to contrasts (pairwise, null, or custom).
}
\details{
The function replaces "exposure_var" in the causal_interventions and
null_intervention with the actual name of the exposure variable.

For contrasts:
- "pairwise": Indicates that all interventions will be compared to each other.
- "null": Indicates that all interventions will be compared to a specified null intervention.
- "custom": Indicates that custom contrasts will be used, which should be defined elsewhere.
}
\examples{
boilerplate_methods_causal_interventions(
  causal_interventions = c("Increase exposure_var", "Do not change exposure_var"),
  exposure_var = "political_conservative",
  contrasts = "null",
  null_intervention = "Do not change exposure_var"
)

}
