% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate_methods_confounding_control.R
\name{boilerplate_methods_confounding_control}
\alias{boilerplate_methods_confounding_control}
\title{Generate Confounding Control Description}
\usage{
boilerplate_methods_confounding_control(
  appendix_ref = "B",
  protocol_url = "https://osf.io/ce4t9/"
)
}
\arguments{
\item{appendix_ref}{Character string specifying the appendix reference for covariates.}

\item{protocol_url}{Character string specifying the URL for the study protocol.}
}
\value{
A character string containing the confounding control description in markdown format.
}
\description{
Generate Confounding Control Description
}
\keyword{internal}
