% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate_methods_eligibility_criteria.R
\name{boilerplate_methods_eligibility_criteria}
\alias{boilerplate_methods_eligibility_criteria}
\title{Generate Eligibility Criteria Description}
\usage{
boilerplate_methods_eligibility_criteria(
  inclusion_criteria,
  exclusion_criteria,
  n_participants,
  baseline_wave
)
}
\arguments{
\item{inclusion_criteria}{Character vector specifying the inclusion criteria.}

\item{exclusion_criteria}{Character vector specifying the exclusion criteria.}

\item{n_participants}{Numeric value indicating the number of participants meeting the criteria.}
}
\value{
A character string containing the eligibility criteria description in markdown format.
}
\description{
Generate Eligibility Criteria Description
}
