% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/boilerplate_methods_identification_assumptions.R
\name{boilerplate_methods_identification_assumptions}
\alias{boilerplate_methods_identification_assumptions}
\title{Generate Identification Assumptions Description}
\usage{
boilerplate_methods_identification_assumptions(exposure_var)
}
\arguments{
\item{exposure_var}{Character string specifying the primary exposure variable.}
}
\value{
A character string containing the identification assumptions description in markdown format.
}
\description{
Generate Identification Assumptions Description
}
\keyword{internal}
