% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate_methods_missing_data.R
\name{boilerplate_methods_missing_data}
\alias{boilerplate_methods_missing_data}
\title{Generate Missing Data Handling Description}
\usage{
boilerplate_methods_missing_data(
  estimators = c("lmtp"),
  grf_appendix = "C",
  baseline_wave,
  exposure_wave,
  outcome_wave
)
}
\arguments{
\item{estimators}{Character vector specifying the statistical estimators used.}

\item{grf_appendix}{Character string specifying the appendix for GRF missing data handling.}
}
\value{
A character string containing the missing data handling description in markdown format.
}
\description{
Generate Missing Data Handling Description
}
\keyword{internal}
