% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate_methods_statistical_estimator.R
\name{boilerplate_methods_statistical_estimator}
\alias{boilerplate_methods_statistical_estimator}
\title{Generate Statistical Estimator Description}
\usage{
boilerplate_methods_statistical_estimator(estimators = "lmtp")
}
\arguments{
\item{estimators}{Character vector specifying the statistical estimators used.}
}
\value{
A character string containing the statistical estimator description in markdown format.
}
\description{
Generate Statistical Estimator Description
}
\keyword{internal}
