% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate_methods_target_population.R
\name{boilerplate_methods_target_population}
\alias{boilerplate_methods_target_population}
\title{Generate Target Population Description}
\usage{
boilerplate_methods_target_population(
  statistical_estimator = "lmtp",
  baseline_wave
)
}
\arguments{
\item{statistical_estimator}{Character string specifying the statistical estimator used.}
}
\value{
A character string containing the target population description in markdown format.
}
\description{
Generate Target Population Description
}
\keyword{internal}
