% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate_methods_variables.R
\name{boilerplate_methods_variables}
\alias{boilerplate_methods_variables}
\title{Generate Variables Section for Methods}
\usage{
boilerplate_methods_variables(exposure_var, outcome_vars, ...)
}
\arguments{
\item{exposure_var}{A character string specifying the name of the exposure variable.}

\item{outcome_vars}{A character vector specifying the names of the outcome variables.}

\item{...}{Additional arguments (not used in this basic version).}
}
\value{
A character string containing the markdown-formatted section on variables.
}
\description{
This function generates a markdown-formatted section describing the variables used in the study.
}
