% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_plot.R
\name{build_margot_plot}
\alias{build_margot_plot}
\title{Build Margot plot}
\usage{
build_margot_plot(
  .data,
  type,
  options,
  include_coefficients,
  standardize_label
)
}
\arguments{
\item{.data}{Prepared data frame}

\item{type}{Type of effect (RD or RR)}

\item{options}{Options list}

\item{include_coefficients}{Whether to include coefficient text}

\item{standardize_label}{Type of standardization label (NZ, US, none)}
}
\value{
ggplot object
}
\description{
Build Margot plot
}
\keyword{internal}
