% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_assess_overlap.R
\name{calculate_overlap_statistics}
\alias{calculate_overlap_statistics}
\title{Calculate overlap statistics}
\usage{
calculate_overlap_statistics(
  W,
  W_hat,
  exposure_name,
  bounds = c(0.05, 0.95),
  good_bounds = c(0.1, 0.9)
)
}
\description{
Calculate overlap statistics
}
\keyword{internal}
