% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_assess_overlap.R
\name{calculate_trimming_summary}
\alias{calculate_trimming_summary}
\title{Calculate trimming summary}
\usage{
calculate_trimming_summary(W_hat, bounds = c(0.05, 0.95))
}
\description{
Calculate trimming summary
}
\keyword{internal}
