% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloured_histogram.R
\name{coloured_histogram}
\alias{coloured_histogram}
\title{Create a Coloured Histogram Highlighting Specific Ranges}
\usage{
coloured_histogram(
  df,
  col_name,
  binwidth = 1,
  scale_min = NULL,
  scale_max = NULL,
  highlight_range = "highest"
)
}
\arguments{
\item{df}{The dataframe containing the data to be plotted.}

\item{col_name}{The name of the column for which the histogram will be generated.}

\item{binwidth}{The width of the bins for the histogram; defaults to 1.}

\item{scale_min}{The minimum value to be used for scaling the histogram. If `NULL`,
the minimum value of `col_name` is used.}

\item{scale_max}{The maximum value to be used for scaling the histogram. If `NULL`,
the maximum value of `col_name` is used.}

\item{highlight_range}{Specifies which range to highlight: "lowest", "highest", or
"both". Defaults to "highest".}
}
\value{
A ggplot object of the histogram with highlighted ranges as specified.
}
\description{
This function generates a histogram with specific ranges highlighted to indicate
the highest and/or lowest values within a unit of the specified limits. It allows
customization of bin width and the range to be highlighted.
}
\examples{
# Assuming df_19 is your dataframe and contains the column 'forgiveness'
graph <- coloured_histogram(
  df = df_19,
  col_name = "forgiveness",
  scale_min = 1,
  scale_max = 7,
  highlight_range = "highest",
  binwidth = .1 # Adjust binwidth as needed
)
print(graph)

}
