% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloured_histogram_quantiles.R
\name{coloured_histogram_quantiles}
\alias{coloured_histogram_quantiles}
\title{Visualize Distribution with Automatically Calculated Quantile Highlights}
\usage{
coloured_histogram_quantiles(df, col_name, n_quantiles, binwidth = 1)
}
\arguments{
\item{df}{Dataframe containing the data to be visualized.}

\item{col_name}{Name of the column to create a histogram for. This column should
contain numeric data.}

\item{n_quantiles}{The number of quantiles to calculate. If n_quantiles is 4, this will
produce quartiles; if 5, quintiles; etc. This should be a positive integer greater than 1.}

\item{binwidth}{Width of the bins for the histogram. Can be adjusted for finer or
coarser resolution of the distribution. Default is 1.}
}
\value{
A ggplot object representing the histogram with highlighted quantiles.
  The plot can be printed or modified further.
}
\description{
This function creates a histogram for a specified column in a dataframe,
highlighting quantiles that are automatically calculated based on the number
of quantiles specified by the user. It draws vertical lines for each quantile
in specified colors, with the title of the plot incorporating the capitalized
column name using \code{tools::toTitleCase()}.
}
\examples{
 \dontrun{
# `df_nz` is the included dataframe with a numeric column 'forgiveness':
df_19 <- dplyr::filter(df_nz, wave == 2019)

quantile_plot <- coloured_histogram_quantiles(
  df = df_19,
  col_name = "forgiveness",
  n_quantiles = 4, # producing quartiles
  binwidth = 0.5 # adjust binwidth as needed
)
}
print(quantile_plot)

}
