% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_rate_cv.R
\name{compare_cv_targets}
\alias{compare_cv_targets}
\title{Compare CV AUTOC and QINI Results}
\usage{
compare_cv_targets(cv_data, flipped_outcomes = NULL)
}
\description{
Compare CV AUTOC and QINI Results
}
\keyword{internal}
