% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_qini_dev.R
\name{compute_baseline_qini}
\alias{compute_baseline_qini}
\title{Compute baseline QINI curve}
\usage{
compute_baseline_qini(
  tau_hat,
  dr_scores,
  weights,
  method = "maq_no_covariates",
  n_bootstrap = 200,
  seed = NULL,
  verbose = TRUE
)
}
\description{
Compute baseline QINI curve
}
\keyword{internal}
