% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_qini_dev.R
\name{compute_gain_summaries}
\alias{compute_gain_summaries}
\title{Compute gain summaries at spend levels}
\usage{
compute_gain_summaries(cate_qini, baseline_qini, spend_levels, outcome)
}
\description{
Compute gain summaries at spend levels
}
\keyword{internal}
