% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_interpret_policy_tree.R
\name{compute_leaf_means}
\alias{compute_leaf_means}
\title{Compute average conditional means within leaves}
\usage{
compute_leaf_means(
  leaf_idx,
  predictions,
  conditional_means,
  act_labels,
  leaf_names,
  use_math_notation = FALSE,
  output_format = "bullet"
)
}
\description{
Compute average conditional means within leaves
}
\keyword{internal}
