% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_policy_tree.R
\name{compute_policy_trees_for_model}
\alias{compute_policy_trees_for_model}
\title{Compute policy trees for a single model}
\usage{
compute_policy_trees_for_model(
  model_result,
  model_name,
  covariates,
  not_missing,
  custom_covariates,
  exclude_covariates,
  covariate_mode,
  compute_depth1,
  compute_depth2,
  train_proportion,
  verbose,
  seed,
  tree_method,
  min_node_size
)
}
\description{
Compute policy trees for a single model
}
\keyword{internal}
