% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_qini_curves.R
\name{compute_qini_curves}
\alias{compute_qini_curves}
\title{Compute Qini Curves for Multi-Arm and Binary Treatments (Deprecated)}
\usage{
compute_qini_curves(tau_hat, Y, W = NULL, W_multi = NULL)
}
\arguments{
\item{tau_hat}{Matrix or vector of estimated treatment effects.}

\item{Y}{Vector of observed outcomes.}

\item{W}{Vector of treatment assignments for binary treatment.}

\item{W_multi}{Factor of treatment assignments for multi-arm treatment.}
}
\value{
A data frame containing Qini curve data for plotting.
}
\description{
DEPRECATED: This function is no longer functional and will be removed in a future version.
Please use `compute_qini_curves_binary()` for binary treatments and
`compute_qini_curves_multi_arm()` for multi-arm treatments instead.
}
\seealso{
\code{\link{compute_qini_curves_binary}} for binary treatments
\code{\link{compute_qini_curves_multi_arm}} for multi-arm treatments
}
\keyword{deprecated}
\keyword{internal}
