% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_causal_forest.R
\name{compute_qini_improved}
\alias{compute_qini_improved}
\title{Compute QINI Curves with Improved DR Scores (Internal)}
\usage{
compute_qini_improved(
  Y,
  W,
  X,
  tau_hat,
  weights = NULL,
  seed = 12345,
  n_bootstrap = 200,
  verbose = TRUE
)
}
\description{
Compute QINI Curves with Improved DR Scores (Internal)
}
\keyword{internal}
