% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_qini_dev.R
\name{compute_qini_metrics}
\alias{compute_qini_metrics}
\title{Compute QINI summary metrics}
\usage{
compute_qini_metrics(cate_qini, baseline_qini, outcome)
}
\description{
Compute QINI summary metrics
}
\keyword{internal}
