% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_policy_tree_bootstrap.R
\name{compute_summary_metrics}
\alias{compute_summary_metrics}
\title{Compute summary metrics across all models}
\usage{
compute_summary_metrics(results, verbose)
}
\description{
Compute summary metrics across all models
}
\keyword{internal}
