% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_formula.R
\name{construct_formula}
\alias{construct_formula}
\title{Construct Formula for Model Fitting}
\usage{
construct_formula(
  Y,
  X = 1,
  baseline_vars,
  continuous_X = FALSE,
  splines = FALSE,
  subclass = NULL
)
}
\arguments{
\item{Y}{The response variable name as a string.}

\item{X}{The treatment or exposure variable name as a string.}

\item{baseline_vars}{Vector of baseline covariate names.}

\item{continuous_X}{Logical, whether X is a continuous variable.}

\item{splines}{Logical, whether to apply spline transformation to X.}
}
\value{
A character string representing the formula for the glm function.
}
\description{
This function constructs a formula string for model fitting based on provided specifications,
including handling of continuous variables and application of splines if required.
}
