% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_policy_tree_bootstrap.R
\name{create_accumulator}
\alias{create_accumulator}
\title{Create accumulator for bootstrap statistics}
\usage{
create_accumulator(vars, n_actions)
}
\description{
Create accumulator for bootstrap statistics
}
\keyword{internal}
