% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_rate_cv.R
\name{create_cv_interpretation}
\alias{create_cv_interpretation}
\title{Create Interpretation for CV Results}
\usage{
create_cv_interpretation(cv_results, alpha, adjust, num_folds, target)
}
\description{
Create Interpretation for CV Results
}
\keyword{internal}
