% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ordered_variable.R
\name{create_ordered_variable}
\alias{create_ordered_variable}
\title{Create Ordered Variable Based on Quantile Breaks with Informative Labels}
\usage{
create_ordered_variable(df, var_name, n_divisions = NULL)
}
\arguments{
\item{df}{A data frame containing the variable to be divided into quantiles.}

\item{var_name}{The name of the variable within the data frame to divide into quantiles.}

\item{n_divisions}{The number of quantile divisions to create.}
}
\value{
The input data frame with an additional column representing the ordered factor variable.
}
\description{
Create Ordered Variable Based on Quantile Breaks with Informative Labels
}
