% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_assess_overlap.R
\name{create_propensity_plot}
\alias{create_propensity_plot}
\title{Create propensity score plot}
\usage{
create_propensity_plot(W, W_hat, exposure_name, theme = "classic")
}
\description{
Create propensity score plot
}
\keyword{internal}
