% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_policy_tree_stability.R
\name{create_stability_accumulator}
\alias{create_stability_accumulator}
\title{Create accumulator for stability statistics}
\usage{
create_stability_accumulator(vars, n_actions)
}
\description{
Create accumulator for stability statistics
}
\keyword{internal}
