% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tau_hat_plot.R
\name{create_tau_hat_plot}
\alias{create_tau_hat_plot}
\title{Create Tau Hat Plot}
\usage{
create_tau_hat_plot(tau_hat, outcome)
}
\arguments{
\item{tau_hat}{A matrix of estimated treatment effects.}

\item{outcome}{A character string specifying the name of the outcome variable.}
}
\value{
A ggplot object representing the distribution of tau hat values.
}
\description{
Creates a histogram plot of tau hat values for each treatment comparison.
}
\keyword{internal}
