% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_node_data_with_positions.R
\name{debug_node_data_with_positions}
\alias{debug_node_data_with_positions}
\title{Calculate Node Data and Positions for Policy Tree Visualisation}
\usage{
debug_node_data_with_positions(policy_tree, node_distance_adjustment = -1.5)
}
\arguments{
\item{policy_tree}{A policy tree object, typically obtained from a causal forest model.}

\item{node_distance_adjustment}{The adjustment for spacing between nodes. Default is -1.5.}
}
\value{
A data frame containing information about each node in the policy tree,
  including its ID, whether it's a leaf node, split variable, split value, action,
  left and right child node IDs, and x and y coordinates for visualization.
}
\description{
This internal function processes a policy tree object and calculates the positions
of nodes for visualisation purposes. It extracts relevant information from the
policy tree and computes x and y coordinates for each node.
}
\keyword{internal}
