% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_plot.R
\name{detect_effect_column}
\alias{detect_effect_column}
\title{Helper function to detect effect column and type}
\usage{
detect_effect_column(data)
}
\description{
Helper function to detect effect column and type
}
\keyword{internal}
