% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/df_nz.R
\docType{data}
\name{df_nz}
\alias{df_nz}
\title{Legacy New Zealand Attitudes and Values Study (NZAVS) Simulated Data}
\format{
A data frame (see `?df_margot_example` for a similar structure)

A data frame with X rows and Y variables:
\describe{
  \item{age}{participant age in years}
  \item{agreeableness}{Score on the agreeableness scale. Includes items such as:
    \describe{
      \item{i.}{I sympathize with others' feelings.}
      \item{ii.}{I am not interested in other people's problems. (reversed)}
      \item{iii.}{I feel others' emotions.}
      \item{iv.}{I am not really interested in others. (reversed)}
    }
  }
  \item{...}{Other variables}
}
}
\source{
Simulated data. Copyright (c) 2024 margot package contributors.
}
\usage{
df_nz

data(df_nz)
}
\description{
`r lifecycle::badge("deprecated")`

This dataset is deprecated and will be removed in a future version.
Please use `fetch_margot_data()` instead.

This dataset contains simulated data from The New Zealand Attitudes and Values Study.
}
\details{
This large dataset (27MB) is being phased out to reduce package size.
Use `fetch_margot_data(version = "v1")` to download the equivalent
dataset from the Open Science Framework.

The dataset is licensed under a custom license. You may use the data for non-commercial purposes with appropriate credit, but redistribution of the data in any form, including modified versions, is not permitted to protect privacy.

The code in the margot package is licensed under the GNU General Public License (GPL) v3.0. You can redistribute it and/or modify it under the terms of the GPL as published by the Free Software Foundation. See <http://www.gnu.org/licenses/>.
}
\examples{
\dontrun{
# old approach (deprecated)
data(df_nz)

# new approach
df <- fetch_margot_data(version = "v1")
}

data(df_nz)
head(df_nz)
}
\seealso{
\code{\link{fetch_margot_data}}
}
\keyword{datasets}
