% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_policy_tree_utils.R
\name{.fastpolicytree_strategy_datatype}
\alias{.fastpolicytree_strategy_datatype}
\title{Return Margot's validated fastpolicytree data representation}
\usage{
.fastpolicytree_strategy_datatype()
}
\description{
Return Margot's validated fastpolicytree data representation
}
\keyword{internal}
