% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_robust_marginal.R
\name{double_robust_marginal}
\alias{double_robust_marginal}
\title{Double Robust Marginal Estimation and Tabulation}
\usage{
double_robust_marginal(
  df,
  Y,
  X,
  baseline_vars,
  treat_0,
  treat_1,
  nsims,
  cores,
  family,
  weights = TRUE,
  continuous_X = FALSE,
  splines = FALSE,
  vcov = "HC2",
  verbose = FALSE,
  delta = 1,
  sd = 1,
  new_name,
  estimand = c("ATE", "ATT"),
  type_causal = c("RR", "RD"),
  type_tab = c("RR", "RD")
)
}
\arguments{
\item{df}{A data frame containing the dataset for analysis.}

\item{Y}{The name of the outcome variable as a string.}

\item{X}{The name of the treatment or exposure variable as a string.}

\item{baseline_vars}{A vector of covariate names included in the model.}

\item{treat_0}{The reference level of the treatment variable.}

\item{treat_1}{The treatment level of the treatment variable.}

\item{nsims}{The number of simulations to run, used in bootstrap or Monte Carlo methods.}

\item{cores}{The number of processor cores to use for parallel computation.}

\item{family}{A description of the error distribution and link function to be used in the model.}

\item{weights}{The name of the weights variable in the data frame, or NULL if no weights are to be used.}

\item{continuous_X}{Logical, indicating whether the treatment variable X is continuous.}

\item{splines}{Logical, indicating whether to use spline functions for continuous variables.}

\item{vcov}{The method to use for variance-covariance estimation.}

\item{verbose}{Logical, indicating whether to print detailed output during computation.}

\item{delta}{The exposure contrast represented by an outcome-mean difference in the E-value calculation.}

\item{sd}{The outcome standard deviation used to standardise an outcome-mean difference in the E-value calculation.}

\item{new_name}{A new name to assign to the tabulated output, typically describing the variable or model.}

\item{estimand}{Specifies the target of the causal inference, such as "ATE" (Average Treatment Effect) or "ATT" (Average Treatment on the Treated).}

\item{type_causal}{The type of effect size (e.g., "RR" for Risk Ratio or "RD" for Risk Difference) to be computed in the causal analysis.}

\item{type_tab}{The type of effect size to be used in the tabulation of results.}
}
\value{
A list containing two elements: `causal_results` with the results from the causal analysis, and
        `tab_results` with the tabulated results including E-values and other statistics.
}
\description{
This function provides a double robust approach for estimating causal effects. It first computes
marginal effects using the `causal_contrast_marginal` function, then processes and tabulates these
results using `tab_engine_marginal`. The function handles both continuous and categorical variables,
and allows specification of the type of effect measure for both causal estimation and tabulation.
}
\examples{
\dontrun{
# Assuming you have a dataset `df_ate` and necessary variables defined
results <- double_robust_marginal(
  df = df_ate,
  Y = "t2_kessler_latent_anxiety_z",
  X = "treatment_var",
  baseline_vars = c("age", "gender"),
  treat_1 = "intervention",
  treat_0 = "control",
  nsims = 200,
  cores = 4,
  family = "gaussian",
  weights = TRUE,
  continuous_X = FALSE,
  splines = FALSE,
  estimand = "ATE",
  type_causal = "RD",
  type_tab = "RD",
  vcov = "HC2",
  new_name = "Test Model Effect",
  delta = 1,
  sd = 1
)
}

}
\keyword{internal}
