% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_qini_diagnostic.R
\name{extract_from_diff_gain}
\alias{extract_from_diff_gain}
\title{Extract estimate and SE from diff_gain string}
\usage{
extract_from_diff_gain(dg)
}
\description{
Extract estimate and SE from diff_gain string
}
\keyword{internal}
