% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_interpret_heterogeneity.R
\name{extract_grf_params}
\alias{extract_grf_params}
\title{Extract GRF parameters from models}
\usage{
extract_grf_params(models)
}
\description{
Extract GRF parameters from models
}
\keyword{internal}
