% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_qini_data}
\alias{extract_qini_data}
\title{Extract Qini Data for Plotting}
\usage{
extract_qini_data(qini_obj, name, max_index, verbose = TRUE)
}
\arguments{
\item{qini_obj}{A Qini object.}

\item{name}{Name of the treatment arm.}

\item{max_index}{Maximum index to extend the curve to.}

\item{verbose}{Logical indicating whether to display detailed messages during execution. Default is TRUE.}
}
\value{
A data frame with extracted Qini data.
}
\description{
Extracts Qini curve data from a Qini object and prepares it for plotting.
}
\keyword{internal}
