% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_policy_tree_bootstrap.R
\name{extract_tree_info}
\alias{extract_tree_info}
\title{Extract essential information from a policy tree}
\usage{
extract_tree_info(tree, var_names)
}
\description{
Extract essential information from a policy tree
}
\keyword{internal}
