% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_minimal_decimals}
\alias{format_minimal_decimals}
\title{Format number with minimal decimal places}
\usage{
format_minimal_decimals(x, max_decimals = NULL)
}
\description{
Format number with minimal decimal places
}
\keyword{internal}
