% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_interpret_heterogeneity.R
\name{format_omnibus_results}
\alias{format_omnibus_results}
\title{Format omnibus test results}
\usage{
format_omnibus_results(model_id, model_name, omnibus_results)
}
\description{
Format omnibus test results
}
\keyword{internal}
