% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_interpret_heterogeneity.R
\name{format_qini_differences}
\alias{format_qini_differences}
\title{Format QINI differences at spend levels}
\usage{
format_qini_differences(model_id, qini_results, spend_levels)
}
\description{
Format QINI differences at spend levels
}
\keyword{internal}
