% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_policy_tree_bootstrap.R
\name{generate_bootstrap_seeds}
\alias{generate_bootstrap_seeds}
\title{Generate bootstrap seeds from metaseed}
\usage{
generate_bootstrap_seeds(metaseed, n_bootstrap)
}
\description{
Generate bootstrap seeds from metaseed
}
\keyword{internal}
