% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_policy_tree_bootstrap.R
\name{get_bootstrap_indices}
\alias{get_bootstrap_indices}
\title{Get bootstrap indices based on variation type}
\usage{
get_bootstrap_indices(
  not_missing,
  vary_type,
  train_prop,
  sample_seed,
  split_seed
)
}
\description{
Get bootstrap indices based on variation type
}
\keyword{internal}
