% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_policy_tree_bootstrap.R
\name{get_consensus_info}
\alias{get_consensus_info}
\title{Get consensus tree information}
\usage{
get_consensus_info(object, model_name, depth = 1)
}
\arguments{
\item{object}{Object of class "margot_bootstrap_policy_tree"}

\item{model_name}{Model name to extract}

\item{depth}{Tree depth (1 or 2)}
}
\value{
List with consensus split information
}
\description{
Get consensus tree information
}
