% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_outcome_transformation_info}
\alias{get_outcome_transformation_info}
\title{Get outcome transformation information for inverse transformation}
\usage{
get_outcome_transformation_info(model_name, original_df)
}
\description{
Get outcome transformation information for inverse transformation
}
\keyword{internal}
