% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{group_results_by_comparison}
\alias{group_results_by_comparison}
\title{Group Results by Comparison}
\usage{
group_results_by_comparison(results_list)
}
\arguments{
\item{results_list}{A list of results from margot_multi_arm_causal_forest.}
}
\value{
A list of grouped results by comparison levels.
}
\description{
Groups the results of multi-arm causal forest analysis by comparison levels.
}
\keyword{internal}
