% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_tab.R
\name{group_tab}
\alias{group_tab}
\title{Group and Annotate Treatment Effect Estimates}
\usage{
group_tab(
  df,
  type = c("RR", "RD"),
  order = c("default", "alphabetical", "custom"),
  custom_order = NULL
)
}
\arguments{
\item{df}{Data frame containing the variables of interest.}

\item{type}{Type of treatment effect to analyze. Expected values are 'RR' for Risk Ratio
and 'RD' for Risk Difference. Defaults to 'RR'.}

\item{order}{Specifies the order in which the outcomes should be arranged. Can be
'default' for descending order of the estimate, 'alphabetical' for alphabetical order by outcome,
or 'custom' for a user-defined order. Default is 'default'.}

\item{custom_order}{A vector of custom ordering for the outcomes, applicable if `order` is set to 'custom'.
This should be a vector containing all outcomes in the desired order.}
}
\value{
A data frame that has been arranged based on the specified order and annotated
with treatment effect estimates, estimate labels, and evidence value annotations.
}
\description{
This function arranges and annotates a data frame based on specified
types of treatment effect estimates (RR or RD). It supports different sorting
options including default descending, alphabetical, and custom order.
}
\examples{
# Example using Risk Ratio (RR) and default sorting
result_df <- group_tab(df = analysis_df, type = 'RR')

# Example using Risk Difference (RD) with alphabetical sorting
result_df <- group_tab(df = analysis_df, type = 'RD', order = 'alphabetical')

# Example using custom sorting order
custom_order <- c("Outcome3", "Outcome1", "Outcome2")
result_df <- group_tab(df = analysis_df, type = 'RR', order = 'custom', custom_order = custom_order)

}
