% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here_read.R
\name{here_read}
\alias{here_read}
\title{Read an R Object from RDS in a Specified Directory}
\usage{
here_read(name, dir_path = NULL, quiet = FALSE)
}
\arguments{
\item{name}{Character string specifying the name of the file to be read (with or without the ".rds" extension).}

\item{dir_path}{Character string specifying the directory path from which the file will be read. If NULL (default), uses `push_mods`.}

\item{quiet}{Logical. If TRUE, suppresses console output. Default is FALSE.}
}
\value{
The R object stored in the file.
}
\description{
Reads an RDS file specified by `name` from a directory defined by `dir_path`
or `push_mods`, returning the R object stored within.
If no .rds file is found, it searches for the file without the .rds extension.
This function uses the `here` package to resolve the path, ensuring that file paths are built in a consistent and platform-independent manner.
}
\details{
If `dir_path` is NULL, the `push_mods` variable must be defined in the user's environment or within the package, pointing to the directory from where files are to be read.
This function will first try to read an .rds file. If not found, it will attempt to read the file without the .rds extension.
}
\examples{
\dontrun{
# Assuming `push_mods` is set in your environment to "~/mydata"
# and you have previously saved an RDS file named "result.rds" in that directory
result <- here_read("result")

# Reading from a custom directory
result <- here_read("result", dir_path = "~/custom_dir")
}

}
