% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_and_combine.R
\name{impute_and_combine}
\alias{impute_and_combine}
\title{Perform multiple imputation on a list of data frames and combine the results}
\usage{
impute_and_combine(
  list_df,
  m = 10,
  exclude_vars = c("t0_sample_frame", "id", "t0_sample_origin_names_combined")
)
}
\arguments{
\item{list_df}{A list containing data frames on which to perform multiple imputation.}

\item{m}{The number of multiple imputations to perform for each data frame.}

\item{exclude_vars}{A vector of variable names to be excluded from the imputation model.}
}
\value{
A data frame that combines all imputed datasets, with unnecessary columns removed
        and row names reset.
}
\description{
This function takes a list of data frames, performs multiple imputation to fill in missing
values using the 'mice' package, and combines the imputed datasets into a single dataset.
The imputations are performed separately for each data frame in the list, and the results
are combined into a 'mids' object, which is then cleaned and returned.
}
\examples{
\dontrun{
# Assuming list_df is a list of data frames with missing values
imputed_data <- impute_and_combine(list_df, m = 5)
print(imputed_data)
}

}
