% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_policy_tree_bootstrap.R
\name{interpret_depth2_only}
\alias{interpret_depth2_only}
\title{Interpret depth-2 only bootstrap results}
\usage{
interpret_depth2_only(
  object,
  model_name,
  outcome_name,
  outcome_label,
  metrics,
  n_bootstrap,
  stability_threshold,
  decimal_places,
  format,
  include_theory = TRUE,
  label_mapping = NULL
)
}
\description{
Interpret depth-2 only bootstrap results
}
\keyword{internal}
