% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_rate_cv.R
\name{interpret_single_cv_target}
\alias{interpret_single_cv_target}
\title{Interpret Single CV Target Results}
\usage{
interpret_single_cv_target(cv_data, target, flipped_outcomes = NULL)
}
\description{
Interpret Single CV Target Results
}
\keyword{internal}
