% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_boilerplate_measures.R
\name{manager_boilerplate_measures}
\alias{manager_boilerplate_measures}
\title{Manage Boilerplate Measures (DEPRECATED)}
\usage{
manager_boilerplate_measures(...)
}
\arguments{
\item{...}{All arguments (ignored)}
}
\value{
A message indicating the function is deprecated.
}
\description{
`r lifecycle::badge("deprecated")`
This function is deprecated. Please use `boilerplate::boilerplate_manage_measures()` instead.
Install the new package with `devtools::install_github("go-bayes/boilerplate")`.
}
\examples{
\dontrun{
# This function is deprecated. Use instead:
# devtools::install_github("go-bayes/boilerplate")
# library(boilerplate)
# boilerplate::boilerplate_manage_measures(...)
}

}
