% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_amelia_to_mice.R
\name{margot_amelia_to_mice}
\alias{margot_amelia_to_mice}
\title{convert an amelia object to a mice object}
\usage{
margot_amelia_to_mice(amelia_output)
}
\arguments{
\item{amelia_output}{an object of class `amelia`, containing imputed datasets from the Amelia package}
}
\value{
a `mids` object compatible with the `mice` package, structured with the original dataset and imputed values
}
\description{
convert an amelia object to a mice object
}
\examples{
# load Amelia package and perform imputation
library(Amelia)
data(africa) # example dataset from Amelia package
amelia_output <- amelia(x = africa, m = 5, idvars = "country") # impute data

# convert amelia object to mice object
mids_obj <- margot_amelia_to_mice(amelia_output)

# verify mids object
print(mids_obj)
}
