% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_debug_qini.R
\name{margot_debug_qini}
\alias{margot_debug_qini}
\title{Debug Qini Curve Data}
\usage{
margot_debug_qini(model_results, model_name)
}
\arguments{
\item{model_results}{Output from margot_causal_forest}

\item{model_name}{Name of the model to debug}
}
\description{
Debug Qini Curve Data
}
