% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_boilerplate.R
\name{margot_get_boilerplate}
\alias{margot_get_boilerplate}
\title{Get boilerplate slot text}
\usage{
margot_get_boilerplate(slot, audience = NULL, context = list(), default = NULL)
}
\description{
Get boilerplate slot text
}
\keyword{internal}
