% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_grf_causal_forest.R
\name{margot_grf_causal_forest}
\alias{margot_grf_causal_forest}
\title{Run Multiple Generalized Random Forest (GRF) Causal Forest Models with Enhanced Features}
\usage{
margot_grf_causal_forest(
  data,
  outcome_vars,
  covariates,
  W,
  weights,
  grf_defaults = list(),
  save_data = FALSE,
  compute_rate = TRUE,
  top_n_vars = 10,
  save_models = FALSE
)
}
\arguments{
\item{data}{A data frame containing all necessary variables.}

\item{outcome_vars}{A character vector of outcome variable names to be modeled.}

\item{covariates}{A matrix of covariates to be used in the GRF models.}

\item{W}{A matrix of treatment assignments.}

\item{weights}{A vector of weights for the observations.}

\item{grf_defaults}{A list of default parameters for the GRF models.}

\item{save_data}{Logical indicating whether to save data, covariates, and weights. Default is FALSE.}

\item{compute_rate}{Logical indicating whether to compute RATE for each model. Default is TRUE.}

\item{top_n_vars}{Integer specifying the number of top variables to use for additional computations. Default is 10.}

\item{save_models}{Logical indicating whether to save the full GRF model objects. Default is FALSE.}
}
\value{
A list containing:
  \item{results}{A list of model results, one for each outcome variable.}
  \item{combined_table}{A data frame combining all custom evaluation tables.}
  \item{outcome_vars}{The character vector of outcome variable names that were modeled.}
  \item{not_missing}{A vector of indices for complete cases.}
  \item{data}{The input data (if save_data is TRUE).}
  \item{covariates}{The input covariates (if save_data is TRUE).}
  \item{weights}{The input weights (if save_data is TRUE).}
  \item{full_models}{A list of full GRF model objects (if save_models is TRUE).}
}
\description{
This function is a wrapper for grf::causal_forest that runs multiple GRF causal forest models
for specified outcome variables. It calculates average treatment effects, tests calibration,
creates custom evaluation tables, and includes additional features such as tau.hat estimates,
RATE calculations, policy trees, variable importance rankings, and best linear projections.
}
