% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_interpret_table.R
\name{margot_interpret_table}
\alias{margot_interpret_table}
\title{Interpret and Describe Causal Effect Estimates}
\usage{
margot_interpret_table(df, causal_scale, estimand)
}
\arguments{
\item{df}{Data frame output from `group_tab` containing categorized causal effect estimates.}

\item{causal_scale}{Character string specifying the causal scale used in the analysis, currently only supports "causal_difference". Future versions may include additional scales.}

\item{estimand}{Character string indicating the type of causal estimand interpreted: "PATE" (Population Average Treatment Effect), "ATE" (Average Treatment Effect), "ATT" (Average Treatment Effect in the Treated), or "CATE" (Conditional Average Treatment Effect).}
}
\value{
A character vector containing a detailed interpretation of each outcome in `df`, including the type of estimand, the causal contrast, E-values, and the strength of evidence for causality.
}
\description{
This function interprets the output of the `group_tab` function, providing textual descriptions of causal effect estimates. It categorizes the strength of evidence for causality based on E-values and confidence intervals, and generates a detailed interpretation of the effect estimates according to specified causal scales and estimands.
}
\examples{
\dontrun{
# Assuming `group_tab_output` is the result from `group_tab`
margot_interpret_table(group_tab_output, "causal_difference", "ATE")
}

}
