% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_merge_databases.R
\name{margot_merge_databases}
\alias{margot_merge_databases}
\title{Merge Two Measure Databases (DEPRECATED, use `boilerplate` package: <https://go-bayes.github.io/boilerplate/>)}
\usage{
margot_merge_databases(...)
}
\arguments{
\item{...}{All arguments (ignored)}
}
\value{
A message indicating the function is deprecated.
}
\description{
`r lifecycle::badge("deprecated")`
This function is deprecated. Please use `boilerplate::boilerplate_merge_databases()` instead.
Install the new package with `devtools::install_github("go-bayes/boilerplate")`.
}
\examples{
\dontrun{
# This function is deprecated. Use instead:
# devtools::install_github("go-bayes/boilerplate")
# library(boilerplate)
# boilerplate::boilerplate_merge_databases(...)
}

}
