% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_omnibus_hetero_test.R
\name{margot_omnibus_hetero_test}
\alias{margot_omnibus_hetero_test}
\title{Omnibus Heterogeneity Test for GRF Models}
\usage{
margot_omnibus_hetero_test(model_results, outcome_vars)
}
\arguments{
\item{model_results}{A list of model results from margot::margot_run_models_grf().}

\item{outcome_vars}{A character vector of outcome variable names.}
}
\value{
A list containing two elements: a data frame with summary statistics and a character vector with interpretations.
}
\description{
This function performs an omnibus heterogeneity test for specified models
outputted from margot::margot_run_models_grf() and provides interpretations for each result.
}
