% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_plot_create_options.R
\name{margot_plot_create_options}
\alias{margot_plot_create_options}
\title{Create Plot Options for Margot Plot}
\usage{
margot_plot_create_options(
  subtitle,
  base_defaults = NULL,
  title = NULL,
  filename_prefix = NULL,
  ...
)
}
\arguments{
\item{subtitle}{Character string. The subtitle for the plot.}

\item{base_defaults}{List. The base default options to use. If not provided, uses a pre-defined set of defaults.}

\item{title}{Character string. The title for the plot. If NULL, uses the title from base_defaults.}

\item{filename_prefix}{Character string. Prefix for the filename. If NULL, uses a default prefix.}

\item{...}{Additional named arguments to override or add to the default options.}
}
\value{
A list of plot options to be used with `margot_plot`.
}
\description{
This function creates a list of options for use with the `margot_plot` function.
It allows for easy customization of plot settings while maintaining a set of default values.
}
\examples{
\dontrun{
health_options <- margot_plot_create_options("Health Outcomes")
education_options <- margot_plot_create_options("Education Outcomes",
                                        title = "Custom Title",
                                        filename_prefix = "edu_outcomes")
}

}
