% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_plot_exposure.R
\name{margot_plot_exposure}
\alias{margot_plot_exposure}
\title{Create a separate exposure plot}
\usage{
margot_plot_exposure(data, name_exposure, exposure_waves, baseline_wave)
}
\arguments{
\item{data}{A data frame containing the exposure data}

\item{name_exposure}{The name of the exposure variable}

\item{exposure_waves}{A vector of wave numbers for exposure measurements}

\item{baseline_wave}{The wave number for baseline measurements}
}
\value{
A ggplot object representing the exposure plot
}
\description{
Create a separate exposure plot
}
