% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_plot_qini.R
\name{margot_plot_qini}
\alias{margot_plot_qini}
\title{Plot Qini Curves from margot_multi_arm_causal_forest Results}
\usage{
margot_plot_qini(mc_result, outcome_var)
}
\arguments{
\item{mc_result}{A list containing the results from margot_multi_arm_causal_forest().}

\item{outcome_var}{A character string specifying the name of the outcome variable
to plot. This should match one of the model names in mc_result$results.}
}
\value{
A ggplot object representing the Qini curves for the specified outcome variable.
}
\description{
This function creates a ggplot object displaying Qini curves based on the
results of a margot_multi_arm_causal_forest() model.
}
\examples{
\dontrun{
# Assuming mc.test is the result of margot_multi_arm_causal_forest()
plot_qini_curves(mc.test, "model_t2_belong_z")
}

}
