% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_plot_categorical.R
\name{margot_plot_shift}
\alias{margot_plot_shift}
\title{Visualise shifts in data distributions with highlighted ranges}
\usage{
margot_plot_shift(
  df,
  col_name,
  label_mapping = NULL,
  binwidth = 1,
  range_highlight = NULL,
  shift = "up",
  show_avg_line = TRUE,
  print_avg_value = TRUE,
  show_sd_line = TRUE,
  title = NULL,
  subtitle = NULL,
  x_lab = NULL,
  y_lab = "Count",
  save_path = NULL,
  width = 12,
  height = 8,
  include_timestamp = FALSE
)
}
\arguments{
\item{label_mapping}{named vector; optional remapping of variable names for labels.}
}
\description{
Visualise shifts in data distributions with highlighted ranges
}
