% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_plot_slope_covariate.R
\name{margot_plot_slope_covariate}
\alias{margot_plot_slope_covariate}
\title{Create a Slope Plot using ggeffects}
\usage{
margot_plot_slope_covariate(
  data,
  formula,
  terms,
  title = NULL,
  y_label = NULL,
  x_label = NULL,
  y_limits = c(1, 7),
  color_label = NULL,
  save_path = NULL,
  width = 12,
  height = 8,
  seed = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame containing the variables for the model.}

\item{formula}{A formula specifying the model to be fit.}

\item{terms}{A character vector specifying the terms to be used in predict_response.}

\item{title}{An optional title for the plot.}

\item{y_label}{An optional label for the y-axis.}

\item{x_label}{An optional label for the x-axis.}

\item{y_limits}{An optional vector of two numbers specifying the y-axis limits. Default is c(1, 7).}

\item{color_label}{An optional label for the color legend.}

\item{save_path}{An optional path to save the plot.}

\item{width}{The width of the saved plot in inches. Default is 12.}

\item{height}{The height of the saved plot in inches. Default is 8.}

\item{seed}{An optional seed for reproducibility.}

\item{...}{Additional arguments to be passed to ggeffects::predict_response.}
}
\value{
A ggplot2 object representing the plot.
}
\description{
This function creates a ggplot2 visualisation using ggeffects to calculate
predicted responses from a model. It allows flexible specification of the model
and plotting options. Draws on the functionality in the `ggeffects` package
}
\examples{
\dontrun{
library(ggplot2)
library(ggeffects)
library(ggokabeito)

# Basic usage
plot <- margot_plot_slope_covariate(
  data = dat,
  formula = warm_muslims ~ wave:hours_work,
  terms = c("wave", "hours_work"),
  title = "Warmth Towards Muslims by Work Hours Over Time",
  y_label = "Warmth",
  x_label = "Wave",
  color_label = "Work Hours"
)

# With additional ggeffects options
plot_with_options <- margot_plot_slope_covariate(
  data = dat,
  formula = warm_immigrants ~ wave:education,
  terms = c("wave", "education [0:18]"),
  title = "Warmth Towards Immigrants by Education Over Time",
  y_label = "Warmth",
  x_label = "Wave",
  color_label = "Education (Years)",
  type = "continuous"
)
}
}
