% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_propensity_model_and_plots.R
\name{margot_propensity_model_and_plots}
\alias{margot_propensity_model_and_plots}
\title{Create Propensity Score Model and Associated Plots}
\usage{
margot_propensity_model_and_plots(
  df_propensity,
  exposure_variable,
  baseline_vars,
  weights_var_name,
  estimand = "ATE",
  method = "ebal",
  focal = NULL,
  love_plot_options = list(),
  bal_tab_options = list(),
  verbose = TRUE
)
}
\arguments{
\item{df_propensity}{A data frame containing the variables for the propensity score model.}

\item{exposure_variable}{A character string specifying the name of the exposure variable in df_propensity.}

\item{baseline_vars}{A character vector specifying the names of the baseline covariates in df_propensity.}

\item{weights_var_name}{A character string specifying the name of the weights variable in df_propensity.}

\item{estimand}{A character string specifying the estimand. Default is "ATE" (Average Treatment Effect).}

\item{method}{A character string specifying the method for propensity score estimation. Default is "ebal".}

\item{focal}{For binary treatments, a value of the treatment to be considered "focal" (i.e., the intervention). Default is NULL.}

\item{love_plot_options}{A list of options to be passed to cobalt::love.plot(). Default is an empty list.}

\item{bal_tab_options}{A list of options to be passed to cobalt::bal.tab(). Default is an empty list.}

\item{verbose}{A logical value indicating whether to print progress messages. Default is TRUE.}
}
\value{
A list containing the following elements:
\itemize{
  \item match_propensity: The propensity score model object.
  \item love_plot: A love plot object created by cobalt::love.plot().
  \item summary: A summary of the propensity score model.
  \item summary_plot: A plot of the propensity score model summary.
  \item balance_table: A balance table created by cobalt::bal.tab().
  \item diagnostics: A list of additional diagnostic information.
}
}
\description{
This function creates a propensity score model using the specified exposure variable
and baseline covariates. It also generates associated plots and diagnostics.
}
\examples{
\dontrun{
# Assuming df_propensity is your dataset with appropriate variables
results <- margot_propensity_model_and_plots(
  df_propensity = df_propensity,
  exposure_variable = "treatment",
  baseline_vars = c("age", "sex", "bmi"),
  weights_var_name = "sample_weights",
  love_plot_options = list(
    thresholds = c(m = .05),
    size = 4
  ),
  bal_tab_options = list(
    thresholds = c(m = .1, v = 2.5)
  ),
  verbose = TRUE
)

# Access the results
print(results$summary)
plot(results$love_plot)
print(results$balance_table)
}

}
