% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_subset_table_grf.R
\name{margot_subset_table_grf}
\alias{margot_subset_table_grf}
\title{Create Combined Tables for Subset Condition Using GRF Models}
\usage{
margot_subset_table_grf(
  models,
  outcome_vars,
  subset_condition = NULL,
  scale = "RD"
)
}
\arguments{
\item{models}{A list of model results from `margot_run_models_grf`.}

\item{outcome_vars}{A character vector of outcome variable names.}

\item{subset_condition}{A logical vector indicating the subset of data to use. Default is NULL.}

\item{scale}{A character string indicating the scale to use in `margot_model_evalue`. Default is "RD".}
}
\value{
A data frame combining all custom evaluation tables for the specified subset condition.
}
\description{
This function calculates new evaluation tables for specified outcome variables
using a subset condition, and combines them into a single data frame. It uses
the models generated by `margot_run_models_grf` to create these tables.
}
\details{
The function uses `margot_model_evalue` to calculate evaluation metrics for each model
in the subset of data specified by `subset_condition`.
}
\note{
Ensure that the `models` list contains results for all specified `outcome_vars`,
otherwise some outcomes may be omitted from the final table.
}
\examples{
\dontrun{
# Assuming result_psychological_well_being is your results object from the previous function
subset_condition <- X[, "t0_self_esteem_z"] < 1
combined_table_subset <- margot_subset_table_grf(
  models = result_psychological_well_being$results,
  outcome_vars = outcomes_psychological_well_being,
  subset_condition = subset_condition,
  scale = "RD"
)
print(combined_table_subset)
}

}
\seealso{
\code{\link{margot_run_models_grf}} for generating the initial models.
}
