% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_resort_contrast_lmtp.R
\name{margot_view_lmtp_structure}
\alias{margot_view_lmtp_structure}
\title{Helper function to view available models and contrasts}
\usage{
margot_view_lmtp_structure(lmtp_output, show = "all")
}
\arguments{
\item{lmtp_output}{Output from margot_lmtp() or margot_lmtp_resort_contrast()}

\item{show}{What to show: "models", "contrasts", "tables", or "all"}
}
\description{
Helper function to view available models and contrasts
}
