% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_wide.R
\name{margot_wide}
\alias{margot_wide}
\title{Transform longitudinal data to wide format with labels}
\usage{
margot_wide(.data, baseline_vars, exposure_var, outcome_vars)
}
\arguments{
\item{.data}{A data frame containing the longitudinal data in long format.}

\item{baseline_vars}{A character vector of baseline variable names to include.}

\item{exposure_var}{A character vector of exposure variable names.}

\item{outcome_vars}{A character vector of outcome variable names.}
}
\value{
A wide-format data frame with prefixed time-stamped columns.
}
\description{
**Deprecated**: Use `margot_wide_machine()` instead.
}
\section{Deprecated}{

This function was soft-deprecated in version 1.0.38 and will be removed in a future release.
}

