% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_wide_impute_baseline.R
\name{margot_wide_impute_baseline}
\alias{margot_wide_impute_baseline}
\title{Transform longitudinal data to wide format and impute baseline (soft-deprecated)}
\usage{
margot_wide_impute_baseline(.data, baseline_vars, exposure_var, outcome_vars)
}
\arguments{
\item{.data}{A data frame containing the longitudinal data in long format.}

\item{baseline_vars}{A character vector of baseline variable names to include and impute.}

\item{exposure_var}{A character vector specifying the names of exposure variables.}

\item{outcome_vars}{A character vector specifying the names of outcome variables.}
}
\value{
A data frame in wide format with imputed baseline variables.
}
\description{
**Deprecated**: Use `margot_wide_machine(imputation_method = "mice")` instead.
}
\section{Deprecated}{

This function was soft-deprecated in version 1.0.38 and will be removed in a future major release.
}

\examples{
# Preferred: use margot_wide_machine with mice imputation
# wide_df <- margot_wide_machine(
#   df,
#   baseline_vars = c("age", "male"),
#   exposure_var = "forgiveness",
#   outcome_vars = "kessler_latent_anxiety",
#   imputation_method = "mice"
# )

}
