% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_plot.R
\name{prepare_margot_plot_data}
\alias{prepare_margot_plot_data}
\title{Prepare data for Margot plot}
\usage{
prepare_margot_plot_data(
  .data,
  type,
  order,
  options,
  label_mapping,
  original_df
)
}
\arguments{
\item{.data}{Input data frame}

\item{type}{Type of effect (RD or RR)}

\item{order}{Ordering method}

\item{options}{Options list}

\item{label_mapping}{Label mapping list}

\item{original_df}{Original data frame for back-transformation}
}
\value{
Prepared data frame
}
\description{
Prepare data for Margot plot
}
\keyword{internal}
