% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_number.R
\name{pretty_number}
\alias{pretty_number}
\title{Format Numbers with Commas}
\usage{
pretty_number(x)
}
\arguments{
\item{x}{A numeric vector that you want to format.}
}
\value{
A character vector where each number is formatted with commas
as thousands separators.
}
\description{
This function takes a numeric vector and formats it by inserting commas as
thousands separators, making large numbers easier to read.
}
\examples{
numbers <- c(1000, 50000, 1234567)
pretty_number(numbers)

}
