% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_causal_forest_dev.R
\name{print.margot_causal_forest_dev}
\alias{print.margot_causal_forest_dev}
\title{Print method for margot_causal_forest_dev objects}
\usage{
\method{print}{margot_causal_forest_dev}(x, ...)
}
\arguments{
\item{x}{A margot_causal_forest_dev object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for margot_causal_forest_dev objects
}
\keyword{internal}
