% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_policy_tree_diagnostics.R
\name{print.margot_correlation_assessment}
\alias{print.margot_correlation_assessment}
\title{Print method for margot_correlation_assessment}
\usage{
\method{print}{margot_correlation_assessment}(x, ...)
}
\description{
Print method for margot_correlation_assessment
}
