% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_rate_cv.R
\name{print.margot_cv_results}
\alias{print.margot_cv_results}
\title{Print method for margot_cv_results}
\usage{
\method{print}{margot_cv_results}(x, ...)
}
\arguments{
\item{x}{A margot_cv_results object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for margot_cv_results
}
\keyword{internal}
