% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_qini_dev.R
\name{print.margot_qini_dev}
\alias{print.margot_qini_dev}
\title{Print method for margot_qini_dev objects}
\usage{
\method{print}{margot_qini_dev}(x, ...)
}
\arguments{
\item{x}{A margot_qini_dev object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for margot_qini_dev objects
}
