% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_naive_regressions.R
\name{process_evalue}
\alias{process_evalue}
\title{Helper function to process E-values for naive regressions}
\usage{
process_evalue(tab, scale, delta, sd)
}
\description{
Helper function to process E-values for naive regressions
}
\keyword{internal}
