% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.R
\name{run_simulations}
\alias{run_simulations}
\title{Run Simulations for Estimating ATE}
\usage{
run_simulations(
  num_simulations,
  N,
  prob_L1,
  A_on_Y,
  L_on_A,
  L_on_Y,
  method = "ps"
)
}
\arguments{
\item{num_simulations}{Integer, number of simulations to run.}

\item{N}{Integer, sample size for each simulation.}

\item{prob_L1}{Double, probability parameter for generating L1.}

\item{A_on_Y}{Double, effect size of A on Y.}

\item{L_on_A}{Double, effect size of L on A.}

\item{L_on_Y}{Double, effect size of L on Y.}

\item{method}{Character, method for IPTW and DR adjustment ("ps" or other).}
}
\value{
A tibble containing the estimated ATEs and their confidence intervals for each method.
}
\description{
This function simulates data and estimates the Average Treatment Effect (ATE)
using different methods under a specified causal model.
}
