% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_plot.R
\name{save_margot_output}
\alias{save_margot_output}
\title{Save Margot plot output}
\usage{
save_margot_output(output, use_timestamp, base_filename, prefix, save_path)
}
\arguments{
\item{output}{Complete output to save}

\item{use_timestamp}{Whether to include timestamp in filename}

\item{base_filename}{Base filename}

\item{prefix}{Optional filename prefix}

\item{save_path}{Directory to save in}
}
\value{
No return value
}
\description{
Save Margot plot output
}
\keyword{internal}
