% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_interpret_heterogeneity.R
\name{select_models}
\alias{select_models}
\title{Select models based on criteria}
\usage{
select_models(
  evidence_summary,
  require_any_positive,
  exclude_negative_any,
  require_omnibus
)
}
\description{
Select models based on criteria
}
\keyword{internal}
