% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_plot.R
\name{setup_margot_plot_options}
\alias{setup_margot_plot_options}
\title{Setup options for Margot plot}
\usage{
setup_margot_plot_options(options, ...)
}
\arguments{
\item{options}{User-provided options list}

\item{...}{Additional parameters to be included in options}
}
\value{
Merged options list
}
\description{
Setup options for Margot plot
}
\keyword{internal}
