% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_qini_dev.R
\name{summary.margot_qini_dev}
\alias{summary.margot_qini_dev}
\title{Summary method for margot_qini_dev objects}
\usage{
\method{summary}{margot_qini_dev}(object, ...)
}
\arguments{
\item{object}{A margot_qini_dev object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for margot_qini_dev objects
}
