% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_policy_tree_bootstrap.R
\name{update_accumulator}
\alias{update_accumulator}
\title{Update accumulator with tree information}
\usage{
update_accumulator(accumulator, tree_info, depth)
}
\description{
Update accumulator with tree information
}
\keyword{internal}
