% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_plot.R
\name{validate_margot_plot_input}
\alias{validate_margot_plot_input}
\title{Validate input for Margot plot}
\usage{
validate_margot_plot_input(.data)
}
\arguments{
\item{.data}{Input data frame}
}
\value{
No return, throws error if invalid
}
\description{
Validate input for Margot plot
}
\keyword{internal}
