% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_recalculate_policy_trees.R
\name{apply_covariate_exclusions}
\alias{apply_covariate_exclusions}
\title{Apply covariate exclusions}
\usage{
apply_covariate_exclusions(covariates, exclude_patterns, verbose = TRUE)
}
\description{
Apply covariate exclusions
}
\keyword{internal}
