% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate_methods.R
\name{boilerplate_methods}
\alias{boilerplate_methods}
\title{Generate Methods Section for a Causal Inference Study (DEPRECATED, use `boilerplate` package: <https://go-bayes.github.io/boilerplate/>)}
\usage{
boilerplate_methods(...)
}
\arguments{
\item{...}{All arguments (ignored)}
}
\value{
A message indicating the function is deprecated.
}
\description{
`r lifecycle::badge("deprecated")`
This function is deprecated. Please use `boilerplate::boilerplate_report_methods()` instead.
Install the new package with `devtools::install_github("go-bayes/boilerplate")`.
}
\examples{
\dontrun{
# This function is deprecated. Use instead:
# devtools::install_github("go-bayes/boilerplate")
# library(boilerplate)
# boilerplate::boilerplate_report_methods(...)
}

}
