% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate_methods_additional_sections.R
\name{boilerplate_methods_additional_sections}
\alias{boilerplate_methods_additional_sections}
\title{Generate Additional Sections for Methods (DEPRECATED, use `boilerplate` package: <https://go-bayes.github.io/boilerplate/>)}
\usage{
boilerplate_methods_additional_sections(...)
}
\arguments{
\item{...}{All arguments (ignored)}
}
\value{
A message indicating the function is deprecated.
}
\description{
`r lifecycle::badge("deprecated")` Use `boilerplate` package: <https://go-bayes.github.io/boilerplate/>

This function is deprecated. Please use `boilerplate::boilerplate_report_additional_sections` instead.
Install the new package with `devtools::install_github("go-bayes/boilerplate")`.
}
\keyword{internal}
