% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate_methods_causal_interventions.R
\name{boilerplate_methods_causal_interventions}
\alias{boilerplate_methods_causal_interventions}
\title{Generate Causal Interventions and Contrasts Section for Methods (DEPRECATED, use `boilerplate` package: <https://go-bayes.github.io/boilerplate/>))}
\usage{
boilerplate_methods_causal_interventions(...)
}
\arguments{
\item{...}{All arguments (ignored)}
}
\value{
A message indicating the function is deprecated.
}
\description{
`r lifecycle::badge("deprecated")`  Use `boilerplate` package: <https://go-bayes.github.io/boilerplate/>
This function is deprecated. Please use `boilerplate::boilerplate_report_causal_interventions()` instead.
Install the new package with `devtools::install_github("go-bayes/boilerplate")`.
}
\examples{
\dontrun{
# This function is deprecated. Use instead:
# devtools::install_github("go-bayes/boilerplate")
# library(boilerplate)
# boilerplate::boilerplate_report_causal_interventions(...)
}

}
\keyword{internal}
