% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate_methods_confounding_control.R
\name{boilerplate_methods_confounding_control}
\alias{boilerplate_methods_confounding_control}
\title{Generate Confounding Control Description (DEPRECATED, use `boilerplate` package: <https://go-bayes.github.io/boilerplate/>)}
\usage{
boilerplate_methods_confounding_control(...)
}
\arguments{
\item{...}{All arguments (ignored)}
}
\value{
A message indicating the function is deprecated.
}
\description{
`r lifecycle::badge("deprecated")`  Use `boilerplate` package: <https://go-bayes.github.io/boilerplate/>
This function is deprecated. Please use `boilerplate::boilerplate_report_confounding_control()` instead.
Install the new package with `devtools::install_github("go-bayes/boilerplate")`.
}
\examples{
\dontrun{
# This function is deprecated. Use instead:
# devtools::install_github("go-bayes/boilerplate")
# library(boilerplate)
# boilerplate::boilerplate_report_confounding_control(...)
}

}
