% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate_methods_eligibility_criteria.R
\name{boilerplate_methods_eligibility_criteria}
\alias{boilerplate_methods_eligibility_criteria}
\title{Generate Eligibility Criteria Description (DEPRECATED)}
\usage{
boilerplate_methods_eligibility_criteria(...)
}
\arguments{
\item{...}{All arguments (ignored)}
}
\value{
A message indicating the function is deprecated.
}
\description{
`r lifecycle::badge("deprecated")`
This function is deprecated. Please use `boilerplate::boilerplate_report_eligibility_criteria()` instead.
Install the new package with `devtools::install_github("go-bayes/boilerplate")`.
}
\examples{
\dontrun{
# This function is deprecated. Use instead:
# devtools::install_github("go-bayes/boilerplate")
# library(boilerplate)
# boilerplate::boilerplate_report_eligibility_criteria(...)
}

}
