% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate_methods_missing_data.R
\name{boilerplate_methods_missing_data}
\alias{boilerplate_methods_missing_data}
\title{Generate Missing Data Handling Description (DEPRECATED)}
\usage{
boilerplate_methods_missing_data(...)
}
\arguments{
\item{...}{All arguments (ignored)}
}
\value{
A message indicating the function is deprecated.
}
\description{
\lifecycle{deprecated}
This function is deprecated. Please use `boilerplate::boilerplate_report_missing_data()` instead.
Install the new package with `devtools::install_github("go-bayes/boilerplate")`.
}
\examples{
\dontrun{
# This function is deprecated. Use instead:
# devtools::install_github("go-bayes/boilerplate")
# library(boilerplate)
# boilerplate::boilerplate_report_missing_data(...)
}

}
