% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate_methods_statistical_estimator.R
\name{boilerplate_methods_statistical_estimator}
\alias{boilerplate_methods_statistical_estimator}
\title{Generate Statistical Estimator Description (DEPRECATED, use `boilerplate` package: <https://go-bayes.github.io/boilerplate/>)}
\usage{
boilerplate_methods_statistical_estimator(...)
}
\arguments{
\item{...}{All arguments (ignored)}
}
\value{
A message indicating the function is deprecated.
}
\description{
`r lifecycle::badge("deprecated")`
This function is deprecated. Please use `boilerplate::boilerplate_report_statistical_estimator()` instead.
Install the new package with `devtools::install_github("go-bayes/boilerplate")`.
}
\examples{
\dontrun{
# This function is deprecated. Use instead:
# devtools::install_github("go-bayes/boilerplate")
# library(boilerplate)
# boilerplate::boilerplate_report_statistical_estimator(...)
}

}
\keyword{internal}
