% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_policy_tree_bootstrap.R
\name{bootstrap_single_model}
\alias{bootstrap_single_model}
\title{Bootstrap analysis for a single model}
\usage{
bootstrap_single_model(
  model_result,
  model_name,
  covariates,
  not_missing,
  custom_covariates,
  exclude_covariates,
  covariate_mode,
  n_bootstrap,
  vary_type,
  depths,
  actual_train_props,
  all_seeds,
  consensus_threshold,
  return_consensus_trees,
  label_mapping,
  verbose,
  seed,
  tree_method
)
}
\description{
Bootstrap analysis for a single model
}
\keyword{internal}
