% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_policy_tree_bootstrap.R,
%   R/margot_policy_tree_stability.R
\name{compute_consensus_info}
\alias{compute_consensus_info}
\title{Compute consensus information from accumulator}
\usage{
compute_consensus_info(accumulator, n_iterations, consensus_threshold)

compute_consensus_info(accumulator, n_iterations, consensus_threshold)
}
\description{
Compute consensus information from accumulator

Compute consensus information from accumulator
}
\keyword{internal}
