% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_qini_curves.R
\name{compute_qini_curves}
\alias{compute_qini_curves}
\title{Compute Qini Curves for Multi-Arm Treatments}
\usage{
compute_qini_curves(tau_hat, Y, W_multi)
}
\arguments{
\item{tau_hat}{Matrix of estimated treatment effects.}

\item{Y}{Vector of observed outcomes.}

\item{W_multi}{Matrix of treatment assignments.}
}
\value{
A data frame containing Qini curve data for plotting.
}
\description{
Computes Qini curves for multi-arm treatments using the maq package.
}
\keyword{internal}
