% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_causal_forest.R
\name{compute_qini_curves_binary}
\alias{compute_qini_curves_binary}
\title{Compute Qini Curves for Binary Treatments}
\usage{
compute_qini_curves_binary(
  tau_hat,
  Y,
  W,
  weights = NULL,
  seed = NULL,
  verbose = TRUE,
  treatment_cost = 1
)
}
\description{
Compute Qini Curves for Binary Treatments
}
\keyword{internal}
