% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_nz.R
\docType{data}
\name{df_nz}
\alias{df_nz}
\title{df_nz: Example Data Frame}
\format{
A data frame with X rows and Y variables:
\describe{
  \item{age}{participant age in years}
  \item{agreeableness}{Score on the agreeableness scale. Includes items such as:
    \describe{
      \item{i.}{I sympathize with others' feelings.}
      \item{ii.}{I am not interested in other people's problems. (reversed)}
      \item{iii.}{I feel others' emotions.}
      \item{iv.}{I am not really interested in others. (reversed)}
    }
  }
  \item{...}{Other variables}
}
}
\source{
Simulated data
}
\usage{
data(df_nz)
}
\description{
This dataset contains simulated data from The New Zealand Attitudes and Values Study.
}
\details{
This dataset is licensed under the Creative Commons Attribution-NonCommercial-NoDerivatives 4.0 International License.
© 2024, Curators of the Margot Package. See LICENSE-data.md for more details.
}
\examples{
data(df_nz)
head(df_nz)
}
\keyword{datasets}
