% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_policy_tree_utils.R
\name{.compute_policy_tree}
\alias{.compute_policy_tree}
\title{Internal wrapper for policy tree computation}
\usage{
.compute_policy_tree(
  X,
  Gamma,
  depth,
  tree_method = "policytree",
  min_node_size = getOption("margot.policy_tree.min_node_size", 1L)
)
}
\arguments{
\item{X}{Covariate matrix}

\item{Gamma}{Matrix of doubly robust scores}

\item{depth}{Integer depth of tree (1 or 2)}

\item{tree_method}{Character string: "policytree" or "fastpolicytree"}
}
\value{
Policy tree object
}
\description{
Internal function that wraps either policytree::policy_tree() or
fastpolicytree::fastpolicytree() depending on user preference and
package availability.
}
\keyword{internal}
