% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_multi_arm_causal_forest.R
\name{group_results_by_comparison}
\alias{group_results_by_comparison}
\title{Group Results by Treatment Comparison}
\usage{
group_results_by_comparison(results)
}
\arguments{
\item{results}{A list of model results.}
}
\value{
A list of data frames, one for each treatment comparison.
}
\description{
Groups custom evaluation tables by treatment comparison level.
}
\keyword{internal}
