% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/margot_multi_arm_causal_forest.R
\name{group_results_by_comparison}
\alias{group_results_by_comparison}
\title{Group Results by Comparison}
\usage{
group_results_by_comparison(results)

group_results_by_comparison(results)
}
\arguments{
\item{results}{A list of model results.}

\item{results_list}{A list of results from margot_multi_arm_causal_forest.}
}
\value{
A list of grouped results by comparison levels.

A list of data frames, one for each treatment comparison.
}
\description{
Groups the results of multi-arm causal forest analysis by comparison levels.

Groups custom evaluation tables by treatment comparison level.
}
\keyword{internal}
