% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_tab.R
\name{group_tab}
\alias{group_tab}
\title{Group and Categorise LMTP Estimates with E-Values}
\usage{
group_tab(df, type = c("RR", "RD"))
}
\arguments{
\item{df}{A dataframe resulting from binding multiple outputs from `margot_lmtp_evalue` or `tab_evalue`.}

\item{type}{A character string specifying the scale of the estimate to be categorised and formatted.
Valid options are "RR" for risk ratio or "RD" for risk difference. Default is "RR".}
}
\value{
A dataframe with estimates categorised as "positive", "negative", or "not reliable" based on their
  confidence intervals and E-values. The dataframe includes a new `estimate_lab` column with formatted
  estimate and confidence interval labels, and an `Estimate` column indicating the category of the estimate.
  Rows are arranged by the size of the estimate and renamed to include the outcome variable.
}
\description{
This function takes a dataframe combining multiple outputs from `margot_lmtp_evalue` or `tab_evalue`,
categorises the estimates based on their confidence intervals and E-values, and prepares the data
for further analysis or plotting. Estimates are categorised as "positive", "negative", or "not reliable".
The function also arranges the data based on the estimate size and adds a column for a formatted label
of the estimate and its confidence interval, including the E-value.
}
\examples{
\dontrun{
# Assuming `tab_all_perceived_support` is a combined dataframe of LMTP estimates
group_tab_all_perceived_support <- group_tab(tab_all_perceived_support)
print(group_tab_all_perceived_support)
}
}
