% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{here_read_arrow}
\alias{here_read_arrow}
\title{Read Data Frame from Parquet File in a Specified Directory (Deprecated)}
\usage{
here_read_arrow(name)
}
\arguments{
\item{name}{Character string specifying the name of the Parquet file to be read.}
}
\description{
This function is deprecated and will be removed in future releases.
For reading data frames, consider using the `here_read_qs` function.
}
\examples{
\dontrun{
my_df <- here_read_arrow("my_dataset")
}

}
\keyword{internal}
