% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here_read_qs.R
\name{here_read_qs}
\alias{here_read_qs}
\title{Read Data Frame from qs File in a Specified Directory}
\usage{
here_read_qs(name, dir_path)
}
\arguments{
\item{name}{Character string specifying the name of the `.qs` file to be read (without the ".qs" extension).}

\item{dir_path}{Character string specifying the directory path from which the file will be read.}
}
\value{
A data frame representing the data stored in the specified `.qs` file.
}
\description{
Reads a `.qs` file specified by `name` from a directory defined by `dir_path`, returning the data frame stored within.
This function uses the `qs` package to efficiently read `.qs` files and the `here` package to construct the file path in a consistent, platform-independent manner.
}
\details{
The `dir_path` argument must point to an existing directory. The function will throw an error if the specified file does not exist or cannot be read as a `.qs` file.
}
