% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here_save_arrow.R
\name{here_save_arrow}
\alias{here_save_arrow}
\title{Save Data Frame to Parquet File in a Specified Directory (Deprecated)}
\usage{
here_save_arrow(df, name)
}
\arguments{
\item{df}{Data frame to be saved.}

\item{name}{Character string specifying the base name of the file.}
}
\description{
This function is deprecated and will be removed in future releases.
For saving data frames, consider using the `here_save_qs` function.
}
\examples{
\dontrun{
my_df <- data.frame(x = 1:5, y = letters[1:5])
here_save_arrow(my_df, "my_saved_dataframe")
}

}
\keyword{internal}
