% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here_save_qs.R
\name{here_save_qs}
\alias{here_save_qs}
\title{Save Data Frame to qs File in a Specified Directory}
\usage{
here_save_qs(df, name, dir_path)
}
\arguments{
\item{df}{Data frame to be saved. This is the object you want to persist to disk in `.qs` format.}

\item{name}{Character string specifying the base name of the file (without the ".qs" extension).}

\item{dir_path}{Character string specifying the directory path where the file will be saved.}
}
\value{
Invisible NULL. The primary effect of this function is the side effect of writing a `.qs` file to disk.
}
\description{
Saves the provided data frame as a `.qs` file using the specified name, within a directory defined by `dir_path`.
This function leverages the `qs` package to write data frames to `.qs` format for efficient storage and quick access in R.
}
\details{
The `dir_path` argument must point to an existing directory. The function does not create directories; it assumes that the specified directory already exists.
}
\examples{
my_df <- data.frame(x = 1:5, y = letters[1:5])
here_save_qs(my_df, "my_saved_dataframe", "~/mydata")

}
