% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_batch_policy.R
\name{margot_batch_policy}
\alias{margot_batch_policy}
\title{Batch Processing of Policy Trees and Related Visualizations}
\usage{
margot_batch_policy(
  result_outcomes,
  policy_tree_args = list(),
  decision_tree_args = list(),
  dpi = 600,
  width = 12,
  height = 12,
  save_plots = TRUE,
  output_dir = here::here(push_mods),
  spend = c(0.2, 0.5)
)
}
\value{
A list where each element corresponds to a model in the input
  `result_outcomes`. Each element is itself a list containing:
  \item{policy_tree_plot}{A ggplot object of the policy tree plot}
  \item{policy_tree_interpretation}{A character string interpreting the policy tree}
  \item{qini_plot}{A ggplot object of the Qini plot}
  \item{decision_tree_visualisation}{A ggplot object visualizing the decision tree}
  \item{policy_combo_plot}{A ggplot object of the policy combo plot}
  \item{diff_gain_summaries}{A nested list containing difference gain summaries for each spend level}
}
\description{
Batch Processing of Policy Trees and Related Visualizations
}
\examples{
\dontrun{
# Example for binary treatment
binary_results <- margot_batch_policy(
  result_outcomes = cf_results,
  policy_tree_args = list(point_alpha = 0.7),
  decision_tree_args = list(text_size = 4),
  dpi = 300,
  width = 12,
  height = 10,
  save_plots = TRUE,
  output_dir = "output_plots_binary",
  spend = c(0.2, 0.5, 0.8)
)

# Access results for a specific binary model and spend level
binary_model_results <- binary_results$model_t2_binary
print(binary_model_results$diff_gain_summaries$spend_0.2$summary)
print(binary_model_results$diff_gain_summaries$spend_0.5$summary)

# Example for multi-arm treatment
multi_arm_results <- margot_batch_policy(
  result_outcomes = mc_results,
  policy_tree_args = list(point_alpha = 0.7),
  decision_tree_args = list(text_size = 4),
  dpi = 300,
  width = 12,
  height = 10,
  save_plots = TRUE,
  output_dir = "output_plots_multi_arm",
  spend = c(0.1, 0.3, 0.5)
)

# Access results for a specific multi-arm model, spend level, and arm
multi_arm_model_results <- multi_arm_results$model_t2_multi_arm
print(multi_arm_model_results$diff_gain_summaries$spend_0.3$all_arms$summary)
print(multi_arm_model_results$diff_gain_summaries$spend_0.5$arm1$summary)

# The plots are saved in the specified output directories
# You can also access the plot objects directly:
print(binary_model_results$policy_tree_plot)
print(multi_arm_model_results$qini_plot)
}

}
