% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_causal_forest.R
\name{margot_causal_forest}
\alias{margot_causal_forest}
\title{Run Multiple Generalized Random Forest (GRF) Causal Forest Models with Enhanced Features}
\usage{
margot_causal_forest(
  data,
  outcome_vars,
  covariates,
  W,
  weights,
  grf_defaults = list(),
  save_data = FALSE,
  compute_rate = TRUE,
  top_n_vars = 15,
  save_models = TRUE,
  train_proportion = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame containing all necessary variables.}

\item{outcome_vars}{A character vector of outcome variable names to be modelled.}

\item{covariates}{A matrix of covariates to be used in the GRF models.}

\item{W}{A vector of binary treatment assignments.}

\item{weights}{A vector of weights for the observations.}

\item{grf_defaults}{A list of default parameters for the GRF models.}

\item{save_data}{Logical indicating whether to save data, covariates, and weights. Default is FALSE.}

\item{compute_rate}{Logical indicating whether to compute RATE for each model. Default is TRUE.}

\item{top_n_vars}{Integer specifying the number of top variables to use for additional computations. Default is 15.}

\item{save_models}{Logical indicating whether to save the full GRF model objects. Default is TRUE.}

\item{train_proportion}{Numeric value between 0 and 1 indicating the proportion of non-missing data to use for
training policy trees. Default is 0.5.}

\item{verbose}{Logical indicating whether to display detailed messages during execution. Default is TRUE.}
}
\value{
A list containing:
  \item{results}{A list of model results, where each element contains model-specific outputs including qini_objects and qini_data}
  \item{combined_table}{A data frame combining custom tables from all models}
  \item{outcome_vars}{The original outcome variables vector}
  \item{not_missing}{Vector of indices for observations with complete data}
  \item{data}{Original data (if save_data=TRUE)}
  \item{covariates}{Original covariates matrix (if save_data=TRUE)}
  \item{weights}{Original weights vector (if save_data=TRUE)}
  \item{full_models}{List of full GRF model objects (if save_models=TRUE)}
}
\description{
This function runs multiple GRF causal forest models with enhanced features. In addition to estimating
causal effects, it can compute the Rank-Weighted Average Treatment Effect (RATE) for each model. RATE
ranks individuals based on their predicted treatment effects and calculates the average effect for selected
percentiles. When computing RATE, the \code{target_rate} parameter determines the variant: \code{"AUTOC"}
(default) exhibits greater power when only a small subset of the population experiences nontrivial heterogeneous
treatment effects, whereas \code{"QINI"} is more powerful when many individuals experience diffuse or substantial
treatment effect heterogeneity.
}
