% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_combine_results.R
\name{margot_combine_results}
\alias{margot_combine_results}
\title{Combine Multiple Results Tables from margot_plot into a Single Formatted Table}
\usage{
margot_combine_results(
  results,
  options = NULL,
  format = "latex",
  digits = 2,
  ...
)
}
\arguments{
\item{results}{A named list of data frames, typically extracted from `margot_plot` objects
using `$transformed_table`. Names will be used as section headers if no options
are provided.}

\item{options}{Optional list of options created by `margot_plot_create_options()`. Should
have the same names as the results list. Each option object can include a 'subtitle'
that will be used as the section header.}

\item{format}{Output format for kable. Default is "latex".}

\item{digits}{Number of decimal places for rounding numeric values. Default is 2.}

\item{...}{Additional arguments passed to kable().}
}
\value{
A kable object that can be further customized using kableExtra functions.
}
\description{
Takes multiple results tables from `margot_plot` objects and combines them into a single
formatted table using kableExtra, with optional group headers for each section.
}
