% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_create_bibliography.R
\name{margot_create_bibliography}
\alias{margot_create_bibliography}
\title{Manage Boilerplate Measures (DEPRECATED, use `boilerplate` package: <https://go-bayes.github.io/boilerplate/>)}
\usage{
margot_create_bibliography(...)
}
\arguments{
\item{...}{All arguments (ignored)}
}
\value{
A message indicating the function is deprecated.
}
\description{
`r lifecycle::badge("deprecated")`
This function is deprecated. Please use `boilerplate::boilerplate_manage_measures()` instead.
Install the new package with `devtools::install_github("go-bayes/boilerplate")`.
}
\examples{
\dontrun{
# This function is deprecated. Use instead:
# devtools::install_github("go-bayes/boilerplate")
# library(boilerplate)
# boilerplate::boilerplate_manage_measures(...)
}

}
\keyword{internal}
