% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_create_database.R
\name{margot_create_database}
\alias{margot_create_database}
\title{Create Database of Measures (DEPRECATED)}
\usage{
margot_create_database(...)
}
\arguments{
\item{...}{All arguments (ignored)}
}
\value{
A message indicating the function is deprecated.
}
\description{
`r lifecycle::badge("deprecated")`
This function is deprecated. Please use `boilerplate::boilerplate_manage_measures()` instead.
Install the new package with `devtools::install_github("go-bayes/boilerplate")`.
}
\examples{
\dontrun{
# This function is deprecated. Use instead:
# devtools::install_github("go-bayes/boilerplate")
# library(boilerplate)
# boilerplate::boilerplate_manage_measures(...)
}

}
