% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_grf_subset_table.R
\name{margot_grf_subset_table}
\alias{margot_grf_subset_table}
\title{Create Combined Tables for Subset Condition Using GRF Models (Deprecated)}
\usage{
margot_grf_subset_table(...)
}
\arguments{
\item{...}{Arguments passed to the deprecated function}
}
\description{
`r lifecycle::badge("deprecated")`
This function is deprecated. Please use `margot_subset_model()` instead.
}
\examples{
\dontrun{
# Instead of using this deprecated function:
# result_subset <- margot_grf_subset_table(...)

# Use the new function:
# result_subset <- margot_subset_model(...)
}
}
