% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_interpret_marginal.R
\name{margot_interpret_marginal}
\alias{margot_interpret_marginal}
\title{Interpret and Describe Causal Effect Estimates Using E-values}
\usage{
margot_interpret_marginal(
  df,
  type = c("RD", "RR"),
  estimand = NULL,
  order = "default"
)
}
\arguments{
\item{df}{Data frame containing causal effect estimates, expected to include columns for
outcome names, effect estimates (either differences or ratios), confidence intervals,
E-values, and a summary estimate label.}

\item{type}{Character string specifying the type of effect estimate. Must be either "RD"
(Risk Difference) or "RR" (Risk Ratio). Default is "RD".}

\item{estimand}{Optional character string indicating the type of causal estimand interpreted: "PATE"
(Population Average Treatment Effect), "ATE" (Average Treatment Effect), "ATT" (Average
Treatment Effect in the Treated), "CATE" (Conditional Average Treatment Effect), or "LMTP"
(Longitudinal Modified Treatment Policy). Default is NULL.}

\item{order}{Character string specifying the order of results. Default is "default".}
}
\value{
A list containing two elements:
  \item{estimand_description}{A character string describing the specified estimand, or NULL if no estimand was provided.}
  \item{interpretation}{A character string containing a detailed interpretation of each outcome in `df`,
  including the causal contrast, E-values, and the strength of evidence for causality.}
}
\description{
This function interprets the output of causal effect analysis, providing textual descriptions
of causal effect estimates. It categorises the strength of evidence for causality based on
E-values and confidence intervals, and generates a detailed interpretation of the effect
estimates according to specified types (i.e., "RD" for risk difference or "RR" for risk ratio)
and estimands.
}
\examples{
\dontrun{
# Assuming `group_tab_output` is the result from a causal analysis
result <- margot_interpret_marginal(group_tab_output, type = "RD", estimand = "ATE")
cat(result$estimand_description)
cat(result$interpretation)

# Using Risk Ratio without specifying an estimand
result <- margot_interpret_marginal(group_tab_output, type = "RR")
cat(result$interpretation)
}

}
