% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_interpret_policy_batch.R
\name{margot_interpret_policy_batch}
\alias{margot_interpret_policy_batch}
\title{Batch Process Policy Tree Interpretations}
\usage{
margot_interpret_policy_batch(
  models,
  save_path = NULL,
  prefix = NULL,
  include_timestamp = FALSE,
  ...
)
}
\arguments{
\item{models}{A list containing the results from multi-arm causal forest models.}

\item{save_path}{The path where the interpretations will be saved. If NULL, interpretations are not saved. Default is NULL.}

\item{prefix}{An optional prefix for the filenames. Default is NULL.}

\item{include_timestamp}{Logical, whether to include a timestamp in the filename. Default is FALSE.}

\item{...}{Additional arguments to pass to margot_interpret_policy_tree().}
}
\value{
A list of the generated interpretations.
}
\description{
This function batch processes outputs from margot_interpret_policy_tree() models
and optionally saves them using the here_save_qs method.
}
