% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_interpret_policy_tree.R
\name{margot_interpret_policy_tree}
\alias{margot_interpret_policy_tree}
\title{Interpret Policy Tree Results}
\usage{
margot_interpret_policy_tree(
  mc_test,
  model_name,
  train_proportion = 0.8,
  custom_action_names = NULL,
  format = "markdown"
)
}
\arguments{
\item{mc_test}{A list containing the results from a multi-arm causal forest model.}

\item{model_name}{A string specifying which model's results to interpret.}

\item{train_proportion}{Numeric value between 0 and 1 for the proportion of data used for training. Default is 0.8.}

\item{custom_action_names}{Optional vector of custom names for the actions. Must match the number of actions in the policy tree.}

\item{format}{A string specifying the output format. Either "markdown" (default) or "latex".}
}
\value{
Invisibly returns a string containing the interpretation in the specified format.
  The function also prints the interpretation to the console.
}
\description{
This function creates an interpretation of policy tree results from a causal forest or multi-arm causal forest model.
It generates a formatted description of the policy tree, including the main splits and recommended actions.
}
\examples{
\dontrun{
# For markdown output (default):
margot_interpret_policy_tree(result_outcomes_health, "model_t2_log_hours_exercise_z")

# For LaTeX output:
margot_interpret_policy_tree(result_outcomes_health, "model_t2_log_hours_exercise_z", format = "latex")

# To store the interpretation in a variable:
interpretation <- margot_interpret_policy_tree(result_outcomes_health, "model_t2_log_hours_exercise_z")
}

}
