% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_rate.R
\name{margot_interpret_rate_comparison}
\alias{margot_interpret_rate_comparison}
\title{Compare and interpret RATE estimates from AUTOC and Qini}
\usage{
margot_interpret_rate_comparison(autoc_df, qini_df, flipped_outcomes = NULL)
}
\arguments{
\item{autoc_df}{Data frame of AUTOC results produced by `margot_rate()`.}

\item{qini_df}{Data frame of Qini  results produced by `margot_rate()`.}

\item{flipped_outcomes}{Character vector of outcomes inverted during preprocessing.}
}
\value{
A list with comparison text, individual summaries, and model name sets.
}
\description{
Compare and interpret RATE estimates from AUTOC and Qini
}
\keyword{internal}
