% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_interpret_table.R
\name{margot_interpret_table}
\alias{margot_interpret_table}
\title{Interpret and Describe Causal Effect Estimates Using E-values}
\usage{
margot_interpret_table(df, causal_scale, estimand = NULL, order = "default")
}
\arguments{
\item{df}{Data frame containing causal effect estimates, expected to include columns for
outcome names, effect estimates (either differences or ratios), confidence intervals,
E-values, and a summary estimate label. The structure of `df` should align with the specified
`causal_scale`.}

\item{causal_scale}{Character string specifying the causal scale used in the analysis.
Currently supports "causal_difference" for differences in means or medians, and "risk_ratio"
for comparing ratios of probabilities or risks.}

\item{estimand}{Optional character string indicating the type of causal estimand interpreted: "PATE"
(Population Average Treatment Effect), "ATE" (Average Treatment Effect), "ATT" (Average
Treatment Effect in the Treated), "CATE" (Conditional Average Treatment Effect), or "LMTP"
(Longitudinal Modified Treatment Policy). Default is NULL.}

\item{order}{Character string specifying the order of results. Default is "default".}
}
\value{
A list containing two elements:
  \item{estimand_description}{A character string describing the specified estimand, or NULL if no estimand was provided.}
  \item{interpretation}{A character string containing a detailed interpretation of each outcome in `df`,
  including the causal contrast, E-values, and the strength of evidence for causality.}
}
\description{
This function interprets the output of causal effect analysis, providing textual descriptions
of causal effect estimates. It categorises the strength of evidence for causality based on
E-values and confidence intervals, and generates a detailed interpretation of the effect
estimates according to specified causal scales (i.e., "causal_difference" or "risk_ratio")
and estimands. This function supports interpreting results on both the causal difference
and risk ratio scales.
}
\examples{
\dontrun{
# Assuming `group_tab_output` is the result from a causal analysis
result <- margot_interpret_table(group_tab_output, "causal_difference", "ATE")
cat(result$estimand_description)
cat(result$interpretation)

# Without specifying an estimand
result <- margot_interpret_table(group_tab_output, "risk_ratio")
cat(result$interpretation)
}

}
