% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_plot.R
\name{margot_plot}
\alias{margot_plot}
\title{Visualise Causal Effect Estimates with E-Values}
\usage{
margot_plot(
  .data,
  type = c("RD", "RR"),
  title,
  subtitle,
  ylab = "",
  estimate_scale = 1,
  base_size = 11,
  text_size = 2.75,
  point_size = 0.5,
  title_size = 10,
  subtitle_size = 9,
  legend_text_size = 6,
  legend_title_size = 6,
  x_offset = ifelse(type == "RR", 0, -1.75),
  x_lim_lo = ifelse(type == "RR", 0.1, -1.75),
  x_lim_hi = ifelse(type == "RR", 2.5, 1),
  linewidth = 0.5,
  simplify_axis = FALSE
)
}
\arguments{
\item{.data}{Data frame containing causal effect estimates outputtd by the `group_tab` function.}

\item{type}{Character string indicating the scale of effect estimates: "RD" for risk difference
or "RR" for risk ratio. Defaults to "RD".}

\item{title}{Main title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{ylab}{Y-axis label, typically left blank as outcomes serve as labels.}

\item{estimate_scale}{Scale factor for estimate label offsets. Default is 1.}

\item{base_size}{Base font size in the plot. Default is 11.}

\item{text_size}{Font size for estimate labels. Default is 2.75.}

\item{point_size}{Size of the points representing estimates. Default is 0.5.}

\item{title_size}{Font size for the plot title. Default is 10.}

\item{subtitle_size}{Font size for the subtitle. Default is 9.}

\item{legend_text_size}{Font size for legend text. Default is 6.}

\item{legend_title_size}{Font size for legend titles. Default is 6.}

\item{x_offset}{Horizontal adjustment of estimate labels, varies by `type`.}

\item{x_lim_lo}{Lower limit of the x-axis, adjusted based on `type`.}

\item{x_lim_hi}{Upper limit of the x-axis, adjusted based on `type`.}

\item{linewidth}{Width of the error bars. Default is 0.5.}

\item{simplify_axis}{Boolean to simplify axis for a cleaner look. Default is FALSE.}
}
\value{
A `ggplot` object visualising causal effect estimates with error bars and categorisation labels.
}
\description{
Creates a graph to visualise multiple causal effect estimates, organised by effect size,
on either risk difference (RD) or risk ratio (RR) scales. It categorises estimates as "positive",
"negative", or "zero_crossing" based on their confidence intervals. This function is flexible,
allowing customisation of graphical parameters and axis simplification.
}
\examples{
\dontrun{
title <- "Religious Service At Least Once Per Week vs None"

# Example using Risk Difference (RD) scale
plot_example_rd <- margot_plot(
  data = your_data_frame, # replace your_data_frame output of group_tab() function
  type = "RD",
  title = title,
  subtitle = "subtitle here",
  estimate_scale = 1,
  linewidth = 0.5,
  simplify_axis = TRUE,
  base_size = 11,
  text_size = 2.75,
  point_size = 0.5,
  title_size = 10,
  subtitle_size = 9,
  legend_text_size = 6,
  legend_title_size = 6,
  x_offset = -1.75,
  x_lim_lo = -1.75,
  x_lim_hi = 1
)
print(plot_example_rd)
}
}
