% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_plot_decision_tree.R
\name{margot_plot_decision_tree}
\alias{margot_plot_decision_tree}
\title{Plot Decision Tree from Policy Tree Results}
\usage{
margot_plot_decision_tree(
  batch_model_output,
  model_name,
  x_padding = 0.25,
  y_padding = 0.05,
  border_size = 1,
  edge_label_hjust_true = 1.5,
  edge_label_hjust_false = -0.5,
  span_ratio = 2,
  text_size = 4,
  node_distance_adjustment = 0.3,
  leaf_fill = "lightgreen",
  non_leaf_fill = "lightyellow",
  title = NULL
)
}
\arguments{
\item{batch_model_output}{A list containing the results from a policy tree model.}

\item{model_name}{A character string specifying the name of the model within
the batch_model_output to plot. This should match one of the model names in
batch_model_output$results.}

\item{x_padding}{The amount of padding to add to the x-axis. Default is 0.25.}

\item{y_padding}{The amount of padding to add to the y-axis. Default is 0.05.}

\item{border_size}{The size of the border around node labels. Default is 1.}

\item{edge_label_hjust_true}{Horizontal justification for "True" edge labels.
Default is 1.5.}

\item{edge_label_hjust_false}{Horizontal justification for "False" edge labels.
Default is -0.5.}

\item{span_ratio}{The aspect ratio of the plot. Default is 2.}

\item{text_size}{The size of the text in the node labels. Default is 4.}

\item{node_distance_adjustment}{The adjustment for spacing between nodes 5 and 6.
Default is 0.3.}

\item{leaf_fill}{The fill color for leaf nodes. Default is "lightgreen".}

\item{non_leaf_fill}{The fill color for non-leaf nodes. Default is "lightyellow".}

\item{title}{The title for the plot. Set to "none" for no title. Default is NULL.}
}
\value{
A ggplot object representing the decision tree for the specified model.
}
\description{
This function creates a ggplot object displaying a decision tree based on the
results of a policy tree model.
}
\examples{
\dontrun{
# Assuming result_psychological_well_being is the batch model output
margot_plot_decision_tree(result_psychological_well_being, "model_t2_hlth_fatigue_z")
}

}
