% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_plot_policy_combo.R
\name{margot_plot_policy_combo}
\alias{margot_plot_policy_combo}
\title{Create a Combined Decision Tree and Policy Relationship Graph}
\usage{
margot_plot_policy_combo(
  mc_test,
  model_name,
  layout = list(heights = c(1, 2)),
  annotation = list(tag_levels = "A")
)
}
\arguments{
\item{mc_test}{An object containing Monte Carlo test results (exact type/class to be specified)}

\item{model_name}{A character string specifying the name of the model}

\item{layout}{A list specifying the layout of the combined plot. Default is
`list(heights = c(1, 2))`, which sets the relative heights of the two plots.}

\item{annotation}{A list specifying the annotation for the combined plot. Default is
`list(tag_levels = "A")`, which adds alphabetic tags to the subplots.}
}
\value{
A list containing:
  \item{plot}{A ggplot2 object representing the combined decision tree and policy relationship graph}
}
\description{
This function generates a combined plot consisting of a decision tree and a graph
showing relationships between variables in the recommended policy.
}
\details{
This function performs the following steps:
1. Generates policy results using `margot_plot_policy_tree`
2. Creates a decision tree plot using `margot_plot_decision_tree`
3. Combines the plots using the patchwork package
4. Applies the specified layout and annotation to the combined plot
}
\examples{
# Assuming mc_test and model_name are already defined:
result <- margot_plot_policy_combo(mc_test, "MyModel")
print(result$plot)

}
\seealso{
\code{\link{margot_plot_policy_tree}}, \code{\link{margot_plot_decision_tree}}
}
