% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_plot_policy_tree.R
\name{margot_plot_policy_tree}
\alias{margot_plot_policy_tree}
\title{Plot Policy Tree Results}
\usage{
margot_plot_policy_tree(
  mc_test,
  model_name,
  color_scale = ggokabeito::scale_colour_okabe_ito(),
  point_alpha = 0.25,
  theme_function = theme_classic,
  title_size = 14,
  subtitle_size = 12,
  axis_title_size = 10,
  legend_title_size = 10,
  jitter_width = 0.3,
  jitter_height = 0.3
)
}
\arguments{
\item{mc_test}{A list containing the results from a multi-arm causal forest model.}

\item{model_name}{A string specifying which model's results to plot.}

\item{color_scale}{A ggplot2 color scale. Default is ggokabeito::scale_colour_okabe_ito().}

\item{point_alpha}{Numeric value between 0 and 1 for point transparency. Default is 0.6.}

\item{theme_function}{A ggplot2 theme function. Default is theme_classic.}

\item{title_size}{Numeric value for the size of the plot title. Default is 14.}

\item{subtitle_size}{Numeric value for the size of the plot subtitle. Default is 12.}

\item{axis_title_size}{Numeric value for the size of axis titles. Default is 10.}

\item{legend_title_size}{Numeric value for the size of the legend title. Default is 10.}

\item{jitter_width}{Numeric value for the amount of horizontal jitter. Default is 0.3.}

\item{jitter_height}{Numeric value for the amount of vertical jitter. Default is 0.3.}
}
\value{
A ggplot object containing two side-by-side plots of the policy tree results.
}
\description{
This function creates a visualization of policy tree results from a causal forest or multi-arm causal forest model.
It generates two plots showing the relationships between the top three split variables,
with points colored by the predicted optimal treatment.
}
\examples{
\dontrun{
plot <- plot_policy_tree_results(mc_test, "model_t2_belong_z")
print(plot)
}

}
