% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_plot_slope_covariate.R
\name{margot_plot_slope_covariate}
\alias{margot_plot_slope_covariate}
\title{Create a Slope Plot using ggeffects}
\usage{
margot_plot_slope_covariate(
  data,
  formula,
  terms,
  title = NULL,
  y_label = NULL,
  x_label = NULL,
  y_limits = c(1, 7),
  color_label = NULL,
  save_path = NULL,
  width = 12,
  height = 8,
  seed = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame containing the variables for the model.}

\item{formula}{A formula specifying the model to be fit.}

\item{terms}{A character vector specifying the terms to be used in predict_response.}

\item{title}{An optional title for the plot. If NULL, an automatic title will be generated.}

\item{y_label}{An optional label for the y-axis. If NULL, the response variable name will be used.}

\item{x_label}{An optional label for the x-axis. If NULL, the first term will be used.}

\item{y_limits}{An optional vector of two numbers specifying the y-axis limits. Default is c(1, 7).}

\item{color_label}{An optional label for the color legend. If NULL, the second term will be used.}

\item{save_path}{An optional path to save the plot. If NULL, the plot will not be saved.}

\item{width}{The width of the saved plot in inches. Default is 12.}

\item{height}{The height of the saved plot in inches. Default is 8.}

\item{seed}{An optional seed for reproducibility.}

\item{...}{Additional arguments to be passed to ggeffects::predict_response.}
}
\value{
A ggplot2 object representing the plot.
}
\description{
This function creates a ggplot2 visualization using ggeffects to calculate
predicted responses from a model. It allows flexible specification of the model
and plotting options. The function automatically handles NA and infinite values,
and reports the number of unique participants and observations used in the analysis.
}
\examples{
\dontrun{
library(ggplot2)
library(ggeffects)
library(ggokabeito)
library(dplyr)

# Basic usage
plot <- margot_plot_slope_covariate(
  data = dat,
  formula = warm_muslims ~ wave:hours_work,
  terms = c("wave", "hours_work"),
  y_label = "Warmth",
  x_label = "Wave",
  color_label = "Work Hours"
)

# With additional ggeffects options
plot_with_options <- margot_plot_slope_covariate(
  data = dat,
  formula = warm_immigrants ~ wave:education,
  terms = c("wave", "education [0:18]"),
  y_label = "Warmth",
  x_label = "Wave",
  color_label = "Education (Years)",
  type = "continuous"
)

# Saving the plot
saved_plot <- margot_plot_slope_covariate(
  data = dat,
  formula = political_orientation ~ wave:age,
  terms = c("wave", "age"),
  save_path = "path/to/save/directory",
  width = 10,
  height = 6
)
}
}
