% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_rate.R
\name{margot_rate}
\alias{margot_rate}
\title{Format RATE Results into Readable Tables}
\usage{
margot_rate(
  models,
  label_mapping = NULL,
  remove_tx_prefix = TRUE,
  remove_z_suffix = TRUE,
  use_title_case = TRUE,
  remove_underscores = TRUE,
  round_digits = 3,
  highlight_significant = TRUE
)
}
\arguments{
\item{models}{A list containing model results from which to extract RATE information.}

\item{label_mapping}{An optional named vector for mapping variable names to readable labels.}

\item{remove_tx_prefix}{Logical indicating whether to remove \sQuote{tx_} prefix from outcome names. Default is TRUE.}

\item{remove_z_suffix}{Logical indicating whether to remove \sQuote{_z} suffix from outcome names. Default is TRUE.}

\item{use_title_case}{Logical indicating whether to convert outcome names to title case. Default is TRUE.}

\item{remove_underscores}{Logical indicating whether to replace underscores with spaces. Default is TRUE.}

\item{round_digits}{Integer specifying the number of decimal places for rounding. Default is 3.}

\item{highlight_significant}{Logical indicating whether to add asterisks to outcomes with positive confidence intervals that don\sQuote{t} cross zero. Default is TRUE}
}
\value{
A list containing two tables: rate_autoc (also accessible as rate_result for backward compatibility) and rate_qini.
}
\description{
Format RATE Results into Readable Tables
}
