% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_run_models_grf.R
\name{margot_run_models_grf}
\alias{margot_run_models_grf}
\title{Run Multiple Generalized Random Forest (GRF) Models with Enhanced Features}
\usage{
margot_run_models_grf(
  data,
  outcome_vars,
  covariates,
  W,
  weights,
  grf_defaults = list(),
  save_data = FALSE,
  compute_rate = TRUE,
  top_n_vars = 10
)
}
\arguments{
\item{data}{A data frame containing all necessary variables.}

\item{outcome_vars}{A character vector of outcome variable names to be modeled.}

\item{covariates}{A matrix of covariates to be used in the GRF models.}

\item{W}{A matrix of treatment assignments.}

\item{weights}{A vector of weights for the observations.}

\item{grf_defaults}{A list of default parameters for the GRF models.}

\item{save_data}{Logical indicating whether to save data, covariates, and weights. Default is FALSE.}

\item{compute_rate}{Logical indicating whether to compute RATE for each model. Default is TRUE.}

\item{top_n_vars}{Integer specifying the number of top variables to use for additional computations. Default is 10.}
}
\value{
A list containing:
  \item{results}{A list of model results, one for each outcome variable.}
  \item{combined_table}{A data frame combining all custom evaluation tables.}
  \item{outcome_vars}{The character vector of outcome variable names that were modeled.}
  \item{tau_hats}{A list of tau.hat estimates for each model.}
  \item{tau_hat_plots}{A list of ggplot objects for tau.hat histograms.}
  \item{rate_results}{A list of RATE results (if compute_rate is TRUE).}
  \item{dr_scores}{A list of double robust scores for each model.}
  \item{policy_trees}{A list of policy trees of depth 1 for each model.}
  \item{not_missing}{A vector of indices for complete cases.}
  \item{variable_importance_rankings}{A list of top n variables by importance for each model, where n is specified by top_n_vars.}
  \item{best_linear_projections}{A list of best linear projection results using top n variables for each model, where n is specified by top_n_vars.}
  \item{policy_trees_depth_2}{A list of policy trees of depth 2 using top n variables for each model, where n is specified by top_n_vars.}
  \item{data}{The input data (if save_data is TRUE).}
  \item{covariates}{The input covariates (if save_data is TRUE).}
  \item{weights}{The input weights (if save_data is TRUE).}
}
\description{
This function runs multiple GRF models for specified outcome variables,
calculates average treatment effects, tests calibration, creates custom
evaluation tables, and includes additional features such as tau.hat estimates,
RATE calculations, policy trees, variable importance rankings, best linear
projections, and depth-2 policy trees.
}
