% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_wide_impute_baseline.R
\name{margot_wide_impute_baseline}
\alias{margot_wide_impute_baseline}
\title{Wide Format Data Transformation with Baseline Imputation}
\usage{
margot_wide_impute_baseline(.data, baseline_vars, exposure_var, outcome_vars)
}
\arguments{
\item{.data}{A dataframe containing the longitudinal data.}

\item{baseline_vars}{A character vector of baseline variable names to include and impute if necessary.}

\item{exposure_var}{A character vector specifying the names of exposure variables.}

\item{outcome_vars}{A character vector specifying the names of outcome variables.}
}
\value{
A data frame in wide format with imputed baseline variables.
}
\description{
This function transforms longitudinal data into a wide format and performs imputation on baseline variables.
It leverages the `mice` package for imputation, ensuring that missing baseline data are appropriately handled.
The function arranges data by participant id and time, filters based on time conditions, and reshapes the data into wide format.
Imputed data are then merged back, maintaining the structure suitable for analysis.
}
\examples{
# the df_nz is a dataset available in the 'margot' package
data(df_nz, package = "margot")

baseline_vars <- c(
  "male", "age", "education_level_coarsen", "eth_cat",
  "partner", "religion_church_binary", "agreeableness",
  "conscientiousness", "extraversion", "honesty_humility",
  "openness", "neuroticism", "sample_weights"
)

exposure_var <- c("forgiveness", "not_lost")

outcome_vars <- c(
  "alcohol_frequency", "alcohol_intensity",
  "hlth_bmi", "hours_exercise"
)

wide_data_imputed <- margot_wide_impute_baseline(df_nz, baseline_vars, exposure_var, outcome_vars)

}
