% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_recalculate_policy_trees.R
\name{recalculate_policy_trees_single}
\alias{recalculate_policy_trees_single}
\title{Recalculate policy trees for a single model}
\usage{
recalculate_policy_trees_single(
  model_result,
  model_name,
  covariates,
  not_missing,
  custom_covariates,
  exclude_covariates,
  covariate_mode,
  verbose,
  seed,
  tree_method
)
}
\description{
Recalculate policy trees for a single model
}
\keyword{internal}
