% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot_policy_tree_bootstrap.R,
%   R/margot_policy_tree_stability.R
\name{reconstruct_consensus_tree}
\alias{reconstruct_consensus_tree}
\title{Reconstruct a consensus tree from consensus splits}
\usage{
reconstruct_consensus_tree(consensus_splits, dr_scores, covariates, depth)

reconstruct_consensus_tree(consensus_splits, dr_scores, covariates, depth)
}
\description{
Reconstruct a consensus tree from consensus splits

Reconstruct a consensus tree from consensus splits
}
\keyword{internal}
